==> R/step0 prepare data/step1 read data.Rmd <==
---
title: "step0 data"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(tidyverse)
library(readxl)
library(summarytools)
```


# rawdata0
new import
```{r}
# rawdata0 <- read_csv('data/Cancer_COVID_project_06182021.csv', col_types = cols(
#   Patient_ID = col_character(),
#   `COVID diag date` = col_date(format='%m/%d/%Y'),
#   BIRTH_YR = col_double(),
#   RACE = col_character(),
#   GENDER = col_character(),
#   ETHNICITY = col_character(),
#   REGION = col_character(),
#   DIVISION = col_character(),
#   DECEASED_INDICATOR = col_double(),
#   DATE_OF_DEATH = col_date(format='%m/%d/%Y'),
#   BMI = col_double(),
#   WT = col_double(),
#   HT = col_double(),
#   SMOKE = col_character(),
#   MA_Cytotoxic_4w = col_double(),
#   PS_Cytotoxic_4w = col_double(),
#   PR_Cytotoxic_4w = col_double(),
#   MA_Non_Cytotoxic_4w = col_double(),
#   PS_Non_Cytotoxic_4w = col_double(),
#   PR_Non_Cytotoxic_4w = col_double(),
#   Cancer_dx_code = col_character(),
#   Cancer_date = col_date(format='%m/%d/%Y'),
#   Cancer = col_double(),
#   Smoke_by_ICD = col_double(),
#   Smoke_by_ICD_group = col_character(),
#   Chronic_Kidney_Disease = col_double(),
#   Chronic_Obstructive_Pulmonary = col_double(),
#   Down_Syndrome = col_double(),
#   Solid_Organ_Transplant = col_double(),
#   Pregnancy = col_double(),
#   Heart_Failure = col_double(),
#   Coronary_Artery_Disease = col_double(),
#   Cardiomyopathies = col_double(),
#   Sickle_Cell_Disease = col_double(),
#   Type_2_Diabetes_Mellitus = col_double(),
#   Asthma = col_double(),
#   Cerebrovascular = col_double(),
#   Cystic_Fibrosis = col_double(),
#   Hypertension = col_double(),
#   Other_Immuno = col_double(),
#   Liver_Disease = col_double(),
#   Dementia = col_double(),
#   Pulmonary_Fibrosis = col_double(),
#   Thalassemia = col_double(),
#   Type_1_Diabetes_Mellitus = col_double(),
#   MA_Hydroxychloroquine = col_double(),
#   MA_Azithromycin = col_double(),
#   PS_Hydroxychloroquine = col_double(),
#   PS_Azithromycin = col_double(),
#   PR_Hydroxychloroquine = col_double(),
#   PR_Azithromycin = col_double(),
#   Oxygen = col_double(),
#   Surgery_4w = col_double(),
#   Radiation_4w = col_double(),
#   covid_admitting_diag = col_double(),
#   Inpatient_visit = col_double(),
#   `discharge disposition` = col_character(),
#   covid_dxcode_during_visit = col_double(),
#   ICU = col_double(),
#   Ventilation = col_double()
# ))
# 
# 
# colnames(rawdata0) <- tolower(colnames(rawdata0))
```


```{r}
rawdata0 <- read_csv('data/Cancer_COVID_project_06182021.csv', col_types = cols(
  # demographic
  Patient_ID = col_character(),
  BIRTH_YR = col_double(),
  RACE = col_character(),
  GENDER = col_character(),
  ETHNICITY = col_character(),
  REGION = col_character(),
  DIVISION = col_character(),
  
  DECEASED_INDICATOR = col_double(),
  DATE_OF_DEATH = col_date(format='%m/%d/%Y'),
  # date
  FIRST_MONTH_ACTIVE = col_double(),
  LAST_MONTH_ACTIVE = col_double(),
  entry_date = col_date(format='%m/%d/%Y'),
  exit_date = col_date(format='%m/%d/%Y'),
  # diagnosis date
  COVID_PCR_date = col_date(format='%m/%d/%Y'),
  # cancer
  cancer = col_double(),
  cancer_dx_codes = col_character(),
  cancer_dx_dates = col_character(),
  solid_tumor = col_double(),
  # comorbidities
  any_malignancy = col_double(),
  hiv = col_double(),
  renal = col_double(),
  hemiplegia = col_double(),
  diabetes_with_cc = col_double(),
  diabetes_without_cc = col_double(),
  severe_liver = col_double(),
  mild_liver = col_double(),
  peptic_ulcer = col_double(),
  rheumatic = col_double(),
  chronic_pulmonary = col_double(),
  cerebrovascular = col_double(),
  dementia = col_double(),
  peripheral_vascular = col_double(),
  heart_failure = col_double(),
  myocardial_infarction = col_double(),
  obesity = col_double(),                    # no need to calculate obesity now
  cardiovascular = col_double(),
  smoking_infarction = col_double(),          # definition of this?
  wheelchair = col_double(),
  dysphagia = col_double(),
  #treatment
  Chemotherapy_DIAG_1_month = col_double(),
  Chemotherapy_PROC_1_month = col_double(),
  Immunotherapy_DIAG_1_month = col_double(),
  Radiation_DIAG_1_month = col_double(),
  Radiation_PROC_1_month = col_double(),
  Chemotherapy_DIAG_3_month = col_double(),
  Chemotherapy_PROC_3_month = col_double(),
  Immunotherapy_DIAG_3_month = col_double(),
  Radiation_DIAG_3_month = col_double(),
  Radiation_PROC_3_month = col_double(),
  Chemotherapy_DIAG_6_month = col_double(),
  Chemotherapy_PROC_6_month = col_double(),
  Immunotherapy_DIAG_6_month = col_double(),
  Radiation_DIAG_6_month = col_double(),
  Radiation_PROC_6_month = col_double(),
  
  # treatment for COVID
  Treatment_PS_Remdesivir = col_double(),
  Treatment_ADM_Remdesivir = col_double(),
  
  Treatment_PS_Lopinavir_or_ritonavir = col_double(),
  Treatment_ADM_Lopinavir_or_ritonavir = col_double(),
  
  
  Treatment_PS_Bamlanivimab = col_double(),
  Treatment_ADM_Bamlanivimab = col_double(),
  Treatment_Bamlanivimab_PROC = col_double(),
  
  Treatment_PS_Casirivimab = col_double(),
  Treatment_ADM_Casirivimab = col_double(),
  Treatment_Casirivimab_PROC = col_double(),
  
  Treatment_PS_Tocilizumab = col_double(),
  Treatment_ADM_Tocilizumab = col_double(),
  Treatment_Tocilizumab_PROC = col_double(),
  
  Treatment_PS_Hydroxychloroquine = col_double(),
  Treatment_ADM_Hydroxychloroquine = col_double(),
  
  
  
  
  # symptom after coronavirus infection
  Fever = col_double(),
  Cough = col_double(),
  Shortness_of_breath = col_double(),
  Muscle = col_double(),
  Headache = col_double(),
  Loss_of_taste_or_smell = col_double(),
  Sore_throat = col_double(),
  Nausea = col_double(),
  Diarrhea = col_double(),
  # outcome
  INTUBATION = col_double(),
  Plasmapharesis = col_double(),
  # covid outcome
  VENTILATOR = col_double(),
  Hospitalization = col_double(),
  ICU = col_double()
)
)

colnames(rawdata0) <- tolower(colnames(rawdata0))


# changes:
# missing covariates: smoking, pregnancy
# no inpatient indicator variables: all patients now are inpatient
```


```{r}
# make a demo sample
# rawdata0 <- rawdata0[sample(nrow(rawdata0), 10000, replace = FALSE),]
```



```{r}
rawdata0_dx_code <- read_csv('data/Cancer_COVID_project_06182021_cancer_dxcodes.csv', col_types = cols(
  PTID = col_character(),
  cancer_date = col_date(format='%m/%d/%Y'),
  cancer_dxcode = col_character())
)

colnames(rawdata0_dx_code) <- tolower(colnames(rawdata0_dx_code))



# get the solid/hematological, metastatic history and first cancer date

rawdata0_dx_code <- rawdata0_dx_code %>%
  mutate(dx2 = substr(cancer_dxcode, 1, 2),
         dx3 = substr(cancer_dxcode, 1, 3)) %>% 
  mutate(hematological = ifelse(dx2 %in% c('C8', 'C9', '20'), 1, 0),
         solid = ifelse(hematological == 0, 1, 0),
         metastatic = ifelse(dx3 %in% c('C7B', 'C77', 'C78', 'C79', '196', '197', '198'), 1, 0)) %>% 
  ungroup()

cancer_code_part1 <- rawdata0_dx_code %>%
  group_by(ptid) %>% 
  transmute(
         ptid,
         solid_history = max(solid),
         hematological_history = max(hematological),
         metastatic_history = max(metastatic), 
         first_cancer_date = min(cancer_date)) %>% 
  distinct()

cancer_code_part2 <- rawdata0_dx_code %>% 
  group_by(ptid) %>% 
  arrange(cancer_date) %>% 
  summarise_all(last) %>% 
  transmute(ptid, solid, hematological, metastatic, 
            last_cancer_date = cancer_date,
            last_cancer_dxcode = cancer_dxcode)
  

cancer_code <- merge(cancer_code_part1, cancer_code_part2, by='ptid')


# merge data
rawdata0 <- merge(rawdata0, cancer_code, by.x = 'patient_id', by.y = 'ptid', all.x = TRUE)
```



```{r}
# rawdata0 <- merge(rawdata0, rawdata0_dx_code_first, by.x = 'patient_id', by.y = 'ptid', all.x = TRUE)
```

```{r}
rawdata0 <- rawdata0 %>% 
  select(-c(cancer_dx_dates, cancer_dx_codes))
```

```{r}
# cancer_13type <- rawdata0 %>%
#   rowwise() %>% 
#   transmute(
#   patient_id,
#   cancer_13type = case_when(
#     regexpr('^C67', last_cancer_dxcode) != -1 || regexpr('^188', last_cancer_dxcode) != -1 ~ 'Bladder',
#     regexpr('^C50', last_cancer_dxcode) != -1 || regexpr('^174', last_cancer_dxcode) != -1 ~ 'Breast',
#     regexpr('^C(18|19|20)', last_cancer_dxcode) != -1 || regexpr('^154', last_cancer_dxcode) != -1 ~ 'Colorectal',
#     regexpr('^C541', last_cancer_dxcode) != -1 || regexpr('^1820', last_cancer_dxcode) != -1 ~ 'Endometrial',
#     regexpr('^C(64|65)', last_cancer_dxcode) != -1  || regexpr('^(1890|1891)', last_cancer_dxcode) != -1 ~ 'Kidney (Renal Cell and Renal Pelvis)',
#     regexpr('^C(91|92|93|94|95)', last_cancer_dxcode) != -1 || regexpr('^(204|205|206|207|208)', last_cancer_dxcode) != -1 ~ 'Leukemia (All Types)',
#     regexpr('^C22', last_cancer_dxcode) != -1 || regexpr('^(1550|1551)', last_cancer_dxcode) != -1 ~ 'Liver and Intrahepatic Bile Duct',
#     regexpr('^C34', last_cancer_dxcode) != -1 || regexpr('^162', last_cancer_dxcode) != -1 ~ 'Lung',
#     regexpr('^C43', last_cancer_dxcode) != -1 || regexpr('^172', last_cancer_dxcode) != -1 ~ 'Melanoma',
#     regexpr('^C(82|83|84|85|86|88)', last_cancer_dxcode) != -1 || regexpr('^(200|202)', last_cancer_dxcode) != -1 ~ 'Non-Hodgkin Lymphoma',
#     regexpr('^C25', last_cancer_dxcode) != -1 || regexpr('^157', last_cancer_dxcode) != -1 ~ 'Pancreatic',
#     regexpr('^C61', last_cancer_dxcode) != -1 || regexpr('^185', last_cancer_dxcode) != -1 ~ 'Prostate',
#     regexpr('^C73', last_cancer_dxcode) != -1 || regexpr('^193', last_cancer_dxcode) != -1 ~ 'Thyroid',
#     !is.na(last_cancer_dxcode) ~ 'Other'
#   )
# )


# save(cancer_13type, file = 'data/cancer_13type.RData')
load('data/cancer_13type.RData')
```

```{r}
rawdata0 <- merge(rawdata0, cancer_13type, by = 'patient_id', all.x = TRUE)
```



```{r}
freq(rawdata0$cancer)
freq(rawdata0$cancer_13type, report.nas = FALSE)
```






## create indicator variable recent_cancer
```{r}
#notes: look-back is a good idea
rawdata0 <- rawdata0 %>% mutate(
  recent_cancer = ifelse(last_cancer_date >= (covid_pcr_date - 365), 1, 0)
)

rawdata0 <- rawdata0  %>% rowwise() %>% 
mutate(cancer_3cat = ifelse(cancer == 0, 'no cancer', 
             ifelse(recent_cancer == 1, 'cancer, recent', 'cancer, no recent'))
       )
```


## sample to test
```{r}
# rawdata0 <- rawdata0[sample(nrow(rawdata0), 50000, replace = FALSE),]
```





<!-- exact the earlies date of cancer record as cancer_date -->
<!-- ```{r} -->
<!-- cancer_dx_dates <- str_split(rawdata0$cancer_dx_dates, pattern = '\\|') -->
<!-- cancer_dx_dates <- lapply(cancer_dx_dates, as.Date, format='%m/%d/%Y') -->
<!-- cancer_dx_dates <- lapply(cancer_dx_dates, sort, decreasing = FALSE) -->
<!-- cancer_dx_dates <- lapply(cancer_dx_dates, function(x) if(length(x) == 0) NA_integer_ else x) -->
<!-- cancer_date <- sapply(cancer_dx_dates, `[[`, 1) -->
<!-- rawdata0$cancer_date <- as.Date(cancer_date, origin ="1970-01-01") -->
<!-- ``` -->




<!-- ```{r} -->
<!-- r <- regexpr('.{10}$', rawdata0$cancer_dx_dates, perl = TRUE) -->
<!-- out <- rep(NA, nrow(rawdata0)) -->
<!-- out[!is.na(rawdata0$cancer_dx_dates)] <- regmatches(rawdata0$cancer_dx_dates, r) -->

<!-- rawdata0$cancer_date <- out -->
<!-- ``` -->


# rename variables to be compatiable with the code (without removing the old)
# check: should we just include inpatient patients?
```{r}
# death=DECEASED_INDICATOR; 
# inpatient=Inpatient_visit;


rawdata0 <- rawdata0 %>% 
  rowwise() %>% 
  mutate(death = deceased_indicator,
         ventilation = ventilator,
         inpatient = hospitalization,
         `covid diag date` = covid_pcr_date)
```



```{r}
rawdata0 <- rawdata0 %>%
  mutate(
          # `covid diag date` = as.Date(`covid diag date`),
          # date_of_death = as.Date(date_of_death),
          # cancer_date = as.Date(cancer_date),
          gender = ifelse(gender=='Unknown', NA, gender),
          region = ifelse(region == 'Other/Unknown', NA, region)
         )
```


```{r}
head(rawdata0)
```




# rawdata1
create a new race variable "race_gp"
```{r}
# data rawdata1;
# set B.rawdata0;
# if ETHNICITY="Not Hispanic" and RACE="Caucasian" then race_gp="Non-hispanic white";
# else if ETHNICITY="Not Hispanic" and RACE="African American" then race_gp="Non-hispanic black";
# else if ETHNICITY="Hispanic" then race_gp="Hispanic";
# else race_gp="Others/Unknown";

# nested if-else to create the new variable base on ethnicity and race
rawdata1 <- rawdata0 %>%
  rowwise() %>% 
  mutate(race_gp = ifelse(ethnicity == 'Not Hispanic' && race == "Caucasian", 'Non-hispanic white', 
                          ifelse(ethnicity == 'Not Hispanic' && race == "African American", 'Non-hispanic black',
                                 ifelse(ethnicity == 'Hispanic', 'Hispanic', NA)
                                 )
                          )
         ) 
```


# no smoking in the new data
<!-- recode smoke status as variable "smoke_status"  -->
<!-- ```{r} -->
<!-- # if smoke in ("Current smoker") then smoke_status="a:current"; -->
<!-- # else if smoke in ("Previously smoked") then smoke_status="b:former"; -->
<!-- # else if smoke in ("N","Never smoked") then smoke_status="c:never"; -->
<!-- # else if smoke in ("Not recorded", "Other smoking status","Not currently smoking","Unknown smoking status","") then smoke_status="d:Other/unknown"; -->


<!-- #  -->
<!-- # level_key <- c('Current smoker'='a:current', 'Previously smoked'='b:former', 'Never smoked'='c:never',  -->
<!-- #                'Not currently smoking'= 'd:Other/unknown', -->
<!-- #                'Other smoking status' = 'd:Other/unknown',  -->
<!-- #                'Unknown smoking status'='d:Other/unknown') -->
<!-- # rawdata1 <- rawdata1 %>% mutate(smoke_status=recode_factor(smoke, !!!level_key, .missing ='d:Other/unknown')) -->



<!-- rawdata1 <- rawdata1 %>% -->
<!--   rowwise() %>%  -->
<!--   mutate(smoke_status = ifelse(smoke == 'Current smoker', 'a:current', -->
<!--                               ifelse(smoke == 'Previously smoked', 'b:former', -->
<!--                                    ifelse(smoke %in% c('N', 'Never smoked'), 'c:never', -->
<!--                                           ifelse(smoke %in% c("Not recorded", "Other smoking status","Not currently smoking","Unknown smoking status"), NA))))) -->

<!-- ``` -->


<!-- imputation smoke status by ICD -->
<!-- ```{r} -->
<!-- # *try to fill in some missing by ICD smoke; -->
<!-- # if smoke in ("Not recorded", "Other smoking status","Not currently smoking","Unknown smoking status","") and Smoke_by_ICD=1 and Smoke_by_ICD_group="current" then smoke_status="a:current"; -->
<!-- # if smoke in ("Not recorded", "Other smoking status","Not currently smoking","Unknown smoking status","") and Smoke_by_ICD=1 and Smoke_by_ICD_group="former" then smoke_status="b:former"; -->

<!-- rawdata1 <- rawdata1 %>% -->
<!--   rowwise() %>%  -->
<!--   # use case_when to update variable to make code more succinct -->
<!--   mutate(smoke_status = case_when( -->
<!--     is.na(smoke_status) && smoke_by_icd == 1 && smoke_by_icd_group == 'current' ~ 'a:current', -->
<!--     is.na(smoke_status) && smoke_by_icd == 1 && smoke_by_icd_group == 'former' ~ 'b:former', -->
<!--     TRUE ~ smoke_status -->
<!--   )) -->

<!--   #  -->
<!--   #  -->
<!--   # mutate(smoke_status = ifelse(is.na(smoke_status) && smoke_by_icd == 1 && smoke_by_icd_group == 'current', "a:current", -->
<!--   #                              ifelse(is.na(smoke_status) && smoke_by_icd == 1 && smoke_by_icd_group == 'former', 'b:former', smoke_status) -->
<!--   #                              ) -->
<!--   #        )  -->
<!-- ``` -->


<!-- ```{r} -->
<!-- # if smoke_status in ("a:current","b:former") then smoke_yn=1; -->
<!-- # else if smoke_status not in ("d:Other/unknown") then smoke_yn=0; -->

<!-- rawdata1 <- rawdata1 %>%  -->
<!--   rowwise() %>%  -->
<!--   mutate(smoke_yn = ifelse(smoke_status %in% c("a:current", "b:former"), 1, 0)) -->
<!-- ``` -->





```{r}
# if death=1 or inpatient=1 or ICU=1 or Ventilation=1 then severity=1;
# else if death~=. and inpatient~=. and ICU~=. and
# Ventilation~=. then severity=0;

rawdata1 <- rawdata1 %>% 
  rowwise() %>% 
  mutate(severity = ifelse((death == 1 || 
                              # inpatient == 1 || 
                              icu == 1 || ventilation == 1), 1, 0))
                          
```


```{r}
# event_count=death+inpatient+ICU+Ventilation;
# COVID_year=year(COVID_diag_date);
# age=(covid_year-birth_yr);

rawdata1 <- rawdata1 %>% 
  rowwise() %>% 
  mutate(event_count = death + 
            inpatient +
           icu + ventilation,
                                covid_year = as.numeric(format(covid_pcr_date, "%Y")), 
                                age = (covid_year - birth_yr))
```



# create age ordinal varaible
```{r}
# if age>=0 and age<18 then age_gp="0-18  ";
# else if age>=18 and age<50 then age_gp="18-50";
# else if age>=50 and age<65 then age_gp="50-65";
# else if age>=65 and age<75 then age_gp="65-75";
# else if age>=75 then age_gp=">=75";
# 
# run;

rawdata1 <- rawdata1 %>% 
  rowwise() %>% 
  mutate(age_gp = case_when(
    age >= 0 && age < 18 ~ '0-18',
    age >= 18 && age < 50 ~ '18-50',
    age >= 50 && age < 65 ~ '50-65',
    age >= 65 && age < 75 ~ '65-75',
    age >= 75 ~ '>= 75'
  ))
  
# rawdata1 <- rawdata1 %>% 
#   rowwise() %>%   
#   mutate(age_gp = ifelse(age >= 0 && age < 18, '0-18',
#          ifelse(age >= 18 && age < 50, '18-50', 
#          ifelse(age >= 50 && age < 65, '50-65',
#          ifelse(age >= 65 && age < 75, '65-75',
#          ifelse(age >= 75, '>= 75', NA)))))
#   )
# 
# freq(rawdata1$age_gp)

```

```{r}
mean(rawdata1$birth_yr, na.rm = TRUE)
```


<!-- check whether the processed data has the same fingerprints as SAS processed data -->
<!-- ```{r} -->
<!-- descr(rawdata1, round.digits = 4, var = severity) -->
<!-- ``` -->

```{r}
save(rawdata1, file = 'data/rawdata1.RData')
```


# skip step2 since no more need to calculate obesity
```{r}
rawdata2 <- rawdata1
save(rawdata2, file = 'data/rawdata2.RData')
```
