==> R/step0 prepare data/step2 calculate obesity.Rmd <==
---
title: "step2 calculate obesity"
output: html_document
---

```{r setup, includeFALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
load('data/rawdata1.Rdata')
```

# rawdata2
```{r}
rawdata2 <- rawdata1 %>%
  rowwise() %>% 
  mutate(sex = ifelse(gender == 'Male', 1, 
                                      ifelse(gender == 'Female', 2, NA)),
                                age_in_month = age * 12)
```

```{r}
summary(rawdata2$sex)
```


```{r}
descr(rawdata2, round.digits = 4, var = c(bmi))
```



<!-- ```{r} -->
<!-- BMI_criterion <- " -->
<!-- sex month_lower_limit month_upper_limit P5 P85 P90 P95 P97 -->
<!--  1	24 24.5 14.73732	 	18.16219	 	18.60948	 	19.33801	 	19.85986 -->
<!--  1 	24.5		25.5			14.71929	 	18.11955	 	18.56111	 	19.2789	 	19.79194 -->
<!--  1 	25.5		26.5			14.68361	 	18.03668	 	18.4673	 	19.16466	 	19.66102 -->
<!--  1 	26.5		27.5			14.64843	 	17.957	 	18.37736	 	19.05567	 	19.53658 -->
<!--  1 	27.5		28.5			14.61379	 	17.88047	 	18.29125	 	18.95187	 	19.41849 -->
<!--  1 	28.5		29.5			14.57969	 	17.80704	 	18.20892	 	18.85317	 	19.30665 -->
<!--  1 	29.5		30.5			14.54615	 	17.73667	 	18.13031	 	18.75949	 	19.20097 -->
<!--  1 	30.5		31.5			14.51319	 	17.66932	 	18.05538	 	18.67078	 	19.10132 -->
<!--  1 	31.5		32.5			14.48084	 	17.60495	 	17.98408	 	18.58695	 	19.00761 -->
<!--  1 	32.5		33.5			14.44909	 	17.54351	 	17.91635	 	18.50792	 	18.91973 -->
<!--  1 	33.5		34.5			14.41798	 	17.48496	 	17.85215	 	18.43363	 	18.83758 -->
<!--  1 	34.5		35.5			14.3875	 	17.42927	 	17.79143	 	18.364	 	18.76106 -->
<!--  1 	35.5		36.5			14.35767	 	17.37639	 	17.73414	 	18.29895	 	18.69006 -->
<!--  1 	36.5		37.5			14.32851	 	17.32627	 	17.68022	 	18.23842	 	18.62449 -->
<!--  1 	37.5		38.5			14.30002	 	17.27889	 	17.62963	 	18.18231	 	18.56425 -->
<!--  1 	38.5		39.5			14.27222	 	17.23419	 	17.58231	 	18.13057	 	18.50924 -->
<!--  1 	39.5		40.5			14.2451	 	17.19213	 	17.5382	 	18.08311	 	18.45938 -->
<!--  1 	40.5		41.5			14.21868	 	17.15266	 	17.49725	 	18.03986	 	18.41456 -->
<!--  1 	41.5		42.5			14.19297	 	17.11575	 	17.45941	 	18.00074	 	18.37469 -->
<!--  1 	42.5		43.5			14.16796	 	17.08135	 	17.42462	 	17.96568	 	18.33969 -->
<!--  1 	43.5		44.5			14.14367	 	17.04941	 	17.39282	 	17.93459	 	18.30947 -->
<!--  1 	44.5		45.5			14.12009	 	17.01988	 	17.36395	 	17.90741	 	18.28393 -->
<!--  1 	45.5		46.5			14.09723	 	16.99272	 	17.33795	 	17.88405	 	18.263 -->
<!--  1 	46.5		47.5			14.07509	 	16.96789	 	17.31477	 	17.86444	 	18.24658 -->
<!--  1 	47.5		48.5			14.05366	 	16.94533	 	17.29434	 	17.8485	 	18.23459 -->
<!--  1 	48.5		49.5			14.03296	 	16.92501	 	17.27661	 	17.83614	 	18.22694 -->
<!--  1 	49.5		50.5			14.01296	 	16.90688	 	17.26151	 	17.8273	 	18.22354 -->
<!--  1 	50.5		51.5			13.99367	 	16.89089	 	17.24899	 	17.82189	 	18.22431 -->
<!--  1 	51.5		52.5			13.97509	 	16.87701	 	17.23899	 	17.81983	 	18.22915 -->
<!--  1 	52.5		53.5			13.95722	 	16.86519	 	17.23145	 	17.82104	 	18.23799 -->
<!--  1 	53.5		54.5			13.94003	 	16.8554	 	17.22632	 	17.82544	 	18.25071 -->
<!--  1 	54.5		55.5			13.92353	 	16.8476	 	17.22354	 	17.83295	 	18.26725 -->
<!--  1 	55.5		56.5			13.90771	 	16.84176	 	17.22306	 	17.84349	 	18.2875 -->
<!--  1 	56.5		57.5			13.89257	 	16.83784	 	17.22483	 	17.85699	 	18.31136 -->
<!--  1 	57.5		58.5			13.87809	 	16.8358	 	17.2288	 	17.87335	 	18.33875 -->
<!--  1 	58.5		59.5			13.86426	 	16.83563	 	17.23493	 	17.89252	 	18.36957 -->
<!--  1 	59.5		60.5			13.85108	 	16.83729	 	17.24315	 	17.9144	 	18.40373 -->
<!--  1 	60.5		61.5			13.83855	 	16.84076	 	17.25344	 	17.93893	 	18.44112 -->
<!--  1 	61.5		62.5			13.82665	 	16.846	 	17.26575	 	17.96602	 	18.48166 -->
<!--  1 	62.5		63.5			13.81537	 	16.853	 	17.28003	 	17.99562	 	18.52525 -->
<!--  1 	63.5		64.5			13.80472	 	16.86173	 	17.29625	 	18.02764	 	18.57179 -->
<!--  1 	64.5		65.5			13.79469	 	16.87217	 	17.31437	 	18.06201	 	18.6212 -->
<!--  1 	65.5		66.5			13.78527	 	16.88428	 	17.33435	 	18.09868	 	18.67337 -->
<!--  1 	66.5		67.5			13.77646	 	16.89805	 	17.35616	 	18.13758	 	18.72823 -->
<!--  1 	67.5		68.5			13.76825	 	16.91346	 	17.37975	 	18.17863	 	18.78569 -->
<!--  1 	68.5		69.5			13.76065	 	16.93048	 	17.4051	 	18.22179	 	18.84564 -->
<!--  1 	69.5		70.5			13.75364	 	16.94909	 	17.43217	 	18.26698	 	18.90802 -->
<!--  1 	70.5		71.5			13.74724	 	16.96925	 	17.46092	 	18.31416	 	18.97273 -->
<!--  1 	71.5		72.5			13.74144	 	16.99096	 	17.49133	 	18.36325	 	19.03969 -->
<!--  1 	72.5		73.5			13.73624	 	17.01418	 	17.52335	 	18.41421	 	19.10882 -->
<!--  1 	73.5		74.5			13.73164	 	17.03888	 	17.55696	 	18.46699	 	19.18005 -->
<!--  1 	74.5		75.5			13.72764	 	17.06505	 	17.59212	 	18.52152	 	19.25329 -->
<!--  1 	75.5		76.5			13.72424	 	17.09265	 	17.6288	 	18.57775	 	19.32847 -->
<!--  1 	76.5		77.5			13.72145	 	17.12166	 	17.66696	 	18.63564	 	19.40551 -->
<!--  1 	77.5		78.5			13.71927	 	17.15206	 	17.70658	 	18.69513	 	19.48434 -->
<!--  1 	78.5		79.5			13.71769	 	17.1838	 	17.74762	 	18.75617	 	19.5649 -->
<!--  1 	79.5		80.5			13.71672	 	17.21688	 	17.79004	 	18.81872	 	19.6471 -->
<!--  1 	80.5		81.5			13.71637	 	17.25126	 	17.83382	 	18.88272	 	19.73089 -->
<!--  1 	81.5		82.5			13.71663	 	17.28691	 	17.87892	 	18.94814	 	19.81619 -->
<!--  1 	82.5		83.5			13.71751	 	17.3238	 	17.92532	 	19.01491	 	19.90294 -->
<!--  1 	83.5		84.5			13.71901	 	17.36192	 	17.97296	 	19.083	 	19.99107 -->
<!--  1 	84.5		85.5			13.72113	 	17.40122	 	18.02183	 	19.15236	 	20.08052 -->
<!--  1 	85.5		86.5			13.72387	 	17.44168	 	18.0719	 	19.22295	 	20.17123 -->
<!--  1 	86.5		87.5			13.72724	 	17.48329	 	18.12312	 	19.29471	 	20.26314 -->
<!--  1 	87.5		88.5			13.73124	 	17.52599	 	18.17548	 	19.36761	 	20.35618 -->
<!--  1 	88.5		89.5			13.73587	 	17.56978	 	18.22893	 	19.44161	 	20.45031 -->
<!--  1 	89.5		90.5			13.74113	 	17.61462	 	18.28344	 	19.51666	 	20.54545 -->
<!--  1 	90.5		91.5			13.74702	 	17.66049	 	18.33899	 	19.59272	 	20.64155 -->
<!--  1 	91.5		92.5			13.75355	 	17.70736	 	18.39554	 	19.66974	 	20.73856 -->
<!--  1 	92.5		93.5			13.76071	 	17.7552	 	18.45306	 	19.74769	 	20.83643 -->
<!--  1 	93.5		94.5			13.76852	 	17.80398	 	18.51152	 	19.82652	 	20.93509 -->
<!--  1 	94.5		95.5			13.77695	 	17.85369	 	18.57089	 	19.9062	 	21.03449 -->
<!--  1 	95.5		96.5			13.78603	 	17.90429	 	18.63115	 	19.98668	 	21.13459 -->
<!--  1 	96.5		97.5			13.79575	 	17.95575	 	18.69225	 	20.06793	 	21.23532 -->
<!--  1 	97.5		98.5			13.8061	 	18.00807	 	18.75418	 	20.1499	 	21.33665 -->
<!--  1 	98.5		99.5			13.8171	 	18.0612	 	18.8169	 	20.23256	 	21.43852 -->
<!--  1 	99.5		100.5			13.82873	 	18.11512	 	18.88038	 	20.31587	 	21.54088 -->
<!--  1 	100.5		101.5			13.84101	 	18.16981	 	18.94459	 	20.39979	 	21.64368 -->
<!--  1 	101.5		102.5			13.85392	 	18.22525	 	19.00952	 	20.48429	 	21.74689 -->
<!--  1 	102.5		103.5			13.86747	 	18.28141	 	19.07512	 	20.56933	 	21.85044 -->
<!--  1 	103.5		104.5			13.88166	 	18.33827	 	19.14137	 	20.65487	 	21.9543 -->
<!--  1 	104.5		105.5			13.89648	 	18.3958	 	19.20825	 	20.74089	 	22.05842 -->
<!--  1 	105.5		106.5			13.91194	 	18.45398	 	19.27573	 	20.82733	 	22.16276 -->
<!--  1 	106.5		107.5			13.92804	 	18.5128	 	19.34378	 	20.91417	 	22.26727 -->
<!--  1 	107.5		108.5			13.94476	 	18.57222	 	19.41238	 	21.00138	 	22.37192 -->
<!--  1 	108.5		109.5			13.96212	 	18.63222	 	19.48149	 	21.08893	 	22.47666 -->
<!--  1 	109.5		110.5			13.9801	 	18.69279	 	19.5511	 	21.17677	 	22.58145 -->
<!--  1 	110.5		111.5			13.99871	 	18.7539	 	19.62118	 	21.26488	 	22.68625 -->
<!--  1 	111.5		112.5			14.01795	 	18.81554	 	19.69171	 	21.35323	 	22.79103 -->
<!--  1 	112.5		113.5			14.0378	 	18.87767	 	19.76266	 	21.44178	 	22.89575 -->
<!--  1 	113.5		114.5			14.05828	 	18.94028	 	19.83401	 	21.53051	 	23.00036 -->
<!--  1 	114.5		115.5			14.07937	 	19.00336	 	19.90573	 	21.61938	 	23.10484 -->
<!--  1 	115.5		116.5			14.10107	 	19.06688	 	19.97781	 	21.70837	 	23.20915 -->
<!--  1 	116.5		117.5			14.12338	 	19.13081	 	20.05021	 	21.79745	 	23.31326 -->
<!--  1 	117.5		118.5			14.1463	 	19.19516	 	20.12292	 	21.88659	 	23.41712 -->
<!--  1 	118.5		119.5			14.16982	 	19.25988	 	20.19592	 	21.97576	 	23.52071 -->
<!--  1 	119.5		120.5			14.19394	 	19.32497	 	20.26919	 	22.06494	 	23.624 -->
<!--  1 	120.5		121.5			14.21866	 	19.39041	 	20.3427	 	22.15409	 	23.72696 -->
<!--  1 	121.5		122.5			14.24396	 	19.45618	 	20.41643	 	22.2432	 	23.82955 -->
<!--  1 	122.5		123.5			14.26985	 	19.52226	 	20.49036	 	22.33224	 	23.93175 -->
<!--  1 	123.5		124.5			14.29633	 	19.58864	 	20.56448	 	22.42118	 	24.03353 -->
<!--  1 	124.5		125.5			14.32338	 	19.6553	 	20.63877	 	22.51	 	24.13486 -->
<!--  1 	125.5		126.5			14.35101	 	19.72222	 	20.7132	 	22.59868	 	24.23571 -->
<!--  1 	126.5		127.5			14.3792	 	19.78938	 	20.78775	 	22.68719	 	24.33606 -->
<!--  1 	127.5		128.5			14.40796	 	19.85678	 	20.86242	 	22.77551	 	24.43589 -->
<!--  1 	128.5		129.5			14.43727	 	19.92439	 	20.93718	 	22.86363	 	24.53516 -->
<!--  1 	129.5		130.5			14.46714	 	19.9922	 	21.01201	 	22.95151	 	24.63386 -->
<!--  1 	130.5		131.5			14.49756	 	20.06019	 	21.0869	 	23.03915	 	24.73197 -->
<!--  1 	131.5		132.5			14.52852	 	20.12835	 	21.16183	 	23.12651	 	24.82945 -->
<!--  1 	132.5		133.5			14.56001	 	20.19667	 	21.23679	 	23.21358	 	24.9263 -->
<!--  1 	133.5		134.5			14.59203	 	20.26514	 	21.31175	 	23.30035	 	25.02249 -->
<!--  1 	134.5		135.5			14.62458	 	20.33373	 	21.38671	 	23.38679	 	25.11801 -->
<!--  1 	135.5		136.5			14.65765	 	20.40243	 	21.46165	 	23.47289	 	25.21283 -->
<!--  1 	136.5		137.5			14.69122	 	20.47124	 	21.53655	 	23.55863	 	25.30693 -->
<!--  1 	137.5		138.5			14.72531	 	20.54013	 	21.61141	 	23.644	 	25.40031 -->
<!--  1 	138.5		139.5			14.75989	 	20.6091	 	21.6862	 	23.72897	 	25.49294 -->
<!--  1 	139.5		140.5			14.79496	 	20.67814	 	21.76091	 	23.81354	 	25.58481 -->
<!--  1 	140.5		141.5			14.83052	 	20.74722	 	21.83554	 	23.89769	 	25.67591 -->
<!--  1 	141.5		142.5			14.86655	 	20.81635	 	21.91006	 	23.98141	 	25.76623 -->
<!--  1 	142.5		143.5			14.90306	 	20.88551	 	21.98447	 	24.06469	 	25.85575 -->
<!--  1 	143.5		144.5			14.94002	 	20.95468	 	22.05876	 	24.1475	 	25.94446 -->
<!--  1 	144.5		145.5			14.97745	 	21.02386	 	22.1329	 	24.22985	 	26.03234 -->
<!--  1 	145.5		146.5			15.01532	 	21.09304	 	22.2069	 	24.31172	 	26.11941 -->
<!--  1 	146.5		147.5			15.05363	 	21.1622	 	22.28075	 	24.3931	 	26.20563 -->
<!--  1 	147.5		148.5			15.09238	 	21.23134	 	22.35442	 	24.47397	 	26.29101 -->
<!--  1 	148.5		149.5			15.13155	 	21.30045	 	22.42791	 	24.55434	 	26.37553 -->
<!--  1 	149.5		150.5			15.17113	 	21.36951	 	22.50122	 	24.6342	 	26.4592 -->
<!--  1 	150.5		151.5			15.21113	 	21.43852	 	22.57433	 	24.71352	 	26.54201 -->
<!--  1 	151.5		152.5			15.25152	 	21.50748	 	22.64724	 	24.79232	 	26.62395 -->
<!--  1 	152.5		153.5			15.2923	 	21.57636	 	22.71993	 	24.87058	 	26.70501 -->
<!--  1 	153.5		154.5			15.33347	 	21.64517	 	22.7924	 	24.94829	 	26.78521 -->
<!--  1 	154.5		155.5			15.37501	 	21.71389	 	22.86465	 	25.02545	 	26.86453 -->
<!--  1 	155.5		156.5			15.41692	 	21.78252	 	22.93666	 	25.10206	 	26.94297 -->
<!--  1 	156.5		157.5			15.45918	 	21.85104	 	23.00842	 	25.17811	 	27.02054 -->
<!--  1 	157.5		158.5			15.50179	 	21.91946	 	23.07994	 	25.2536	 	27.09724 -->
<!--  1 	158.5		159.5			15.54474	 	21.98777	 	23.15121	 	25.32853	 	27.17307 -->
<!--  1 	159.5		160.5			15.58801	 	22.05596	 	23.22221	 	25.40289	 	27.24802 -->
<!--  1 	160.5		161.5			15.63161	 	22.12402	 	23.29295	 	25.47668	 	27.32211 -->
<!--  1 	161.5		162.5			15.67551	 	22.19194	 	23.36342	 	25.5499	 	27.39534 -->
<!--  1 	162.5		163.5			15.71971	 	22.25973	 	23.43362	 	25.62256	 	27.46771 -->
<!--  1 	163.5		164.5			15.7642	 	22.32737	 	23.50354	 	25.69464	 	27.53924 -->
<!--  1 	164.5		165.5			15.80897	 	22.39487	 	23.57318	 	25.76616	 	27.60992 -->
<!--  1 	165.5		166.5			15.85401	 	22.46221	 	23.64253	 	25.83712	 	27.67977 -->
<!--  1 	166.5		167.5			15.89931	 	22.52939	 	23.7116	 	25.90751	 	27.74879 -->
<!--  1 	167.5		168.5			15.94486	 	22.5964	 	23.78038	 	25.97734	 	27.817 -->
<!--  1 	168.5		169.5			15.99065	 	22.66325	 	23.84887	 	26.04662	 	27.88441 -->
<!--  1 	169.5		170.5			16.03667	 	22.72993	 	23.91706	 	26.11535	 	27.95102 -->
<!--  1 	170.5		171.5			16.0829	 	22.79643	 	23.98496	 	26.18353	 	28.01686 -->
<!--  1 	171.5		172.5			16.12934	 	22.86275	 	24.05257	 	26.25117	 	28.08193 -->
<!--  1 	172.5		173.5			16.17598	 	22.92889	 	24.11987	 	26.31828	 	28.14624 -->
<!--  1 	173.5		174.5			16.2228	 	22.99485	 	24.18689	 	26.38485	 	28.20983 -->
<!--  1 	174.5		175.5			16.2698	 	23.06062	 	24.25361	 	26.45091	 	28.27269 -->
<!--  1 	175.5		176.5			16.31696	 	23.12619	 	24.32003	 	26.51646	 	28.33484 -->
<!--  1 	176.5		177.5			16.36427	 	23.19158	 	24.38616	 	26.58151	 	28.39632 -->
<!--  1 	177.5		178.5			16.41172	 	23.25677	 	24.452	 	26.64606	 	28.45712 -->
<!--  1 	178.5		179.5			16.4593	 	23.32177	 	24.51755	 	26.71014	 	28.51728 -->
<!--  1 	179.5		180.5			16.507	 	23.38657	 	24.58281	 	26.77374	 	28.57682 -->
<!--  1 	180.5		181.5			16.55481	 	23.45117	 	24.64778	 	26.83688	 	28.63575 -->
<!--  1 	181.5		182.5			16.60271	 	23.51557	 	24.71247	 	26.89958	 	28.6941 -->
<!--  1 	182.5		183.5			16.6507	 	23.57978	 	24.77688	 	26.96184	 	28.75189 -->
<!--  1 	183.5		184.5			16.69875	 	23.64378	 	24.84102	 	27.02368	 	28.80915 -->
<!--  1 	184.5		185.5			16.74687	 	23.70758	 	24.90489	 	27.08511	 	28.8659 -->
<!--  1 	185.5		186.5			16.79503	 	23.77119	 	24.96848	 	27.14616	 	28.92217 -->
<!--  1 	186.5		187.5			16.84323	 	23.83459	 	25.03182	 	27.20683	 	28.97798 -->
<!--  1 	187.5		188.5			16.89146	 	23.89779	 	25.0949	 	27.26714	 	29.03337 -->
<!--  1 	188.5		189.5			16.93969	 	23.9608	 	25.15773	 	27.3271	 	29.08835 -->
<!--  1 	189.5		190.5			16.98792	 	24.02361	 	25.22032	 	27.38675	 	29.14297 -->
<!--  1 	190.5		191.5			17.03615	 	24.08622	 	25.28267	 	27.44609	 	29.19725 -->
<!--  1 	191.5		192.5			17.08434	 	24.14864	 	25.34478	 	27.50514	 	29.25123 -->
<!--  1 	192.5		193.5			17.1325	 	24.21087	 	25.40668	 	27.56393	 	29.30493 -->
<!--  1 	193.5		194.5			17.18061	 	24.27291	 	25.46835	 	27.62247	 	29.35838 -->
<!--  1 	194.5		195.5			17.22865	 	24.33476	 	25.52982	 	27.68078	 	29.41164 -->
<!--  1 	195.5		196.5			17.27662	 	24.39642	 	25.59109	 	27.7389	 	29.46472 -->
<!--  1 	196.5		197.5			17.3245	 	24.4579	 	25.65217	 	27.79683	 	29.51767 -->
<!--  1 	197.5		198.5			17.37229	 	24.5192	 	25.71306	 	27.85461	 	29.57051 -->
<!--  1 	198.5		199.5			17.41995	 	24.58033	 	25.77379	 	27.91225	 	29.6233 -->
<!--  1 	199.5		200.5			17.46749	 	24.64128	 	25.83436	 	27.96979	 	29.67606 -->
<!--  1 	200.5		201.5			17.51489	 	24.70207	 	25.89477	 	28.02724	 	29.72885 -->
<!--  1 	201.5		202.5			17.56214	 	24.76269	 	25.95504	 	28.08464	 	29.78169 -->
<!--  1 	202.5		203.5			17.60923	 	24.82315	 	26.01519	 	28.142	 	29.83463 -->
<!--  1 	203.5		204.5			17.65613	 	24.88346	 	26.07522	 	28.19937	 	29.88771 -->
<!--  1 	204.5		205.5			17.70284	 	24.94362	 	26.13515	 	28.25676	 	29.94097 -->
<!--  1 	205.5		206.5			17.74935	 	25.00363	 	26.19498	 	28.3142	 	29.99447 -->
<!--  1 	206.5		207.5			17.79564	 	25.0635	 	26.25474	 	28.37173	 	30.04824 -->
<!--  1 	207.5		208.5			17.8417	 	25.12324	 	26.31443	 	28.42937	 	30.10232 -->
<!--  1 	208.5		209.5			17.88751	 	25.18286	 	26.37407	 	28.48716	 	30.15677 -->
<!--  1 	209.5		210.5			17.93306	 	25.24235	 	26.43368	 	28.54513	 	30.21164 -->
<!--  1 	210.5		211.5			17.97834	 	25.30173	 	26.49326	 	28.6033	 	30.26696 -->
<!--  1 	211.5		212.5			18.02333	 	25.361	 	26.55284	 	28.66171	 	30.3228 -->
<!--  1 	212.5		213.5			18.06802	 	25.42017	 	26.61243	 	28.72041	 	30.3792 -->
<!--  1 	213.5		214.5			18.11239	 	25.47925	 	26.67204	 	28.77941	 	30.4362 -->
<!--  1 	214.5		215.5			18.15644	 	25.53824	 	26.73169	 	28.83875	 	30.49387 -->
<!--  1 	215.5		216.5			18.20014	 	25.59716	 	26.79141	 	28.89848	 	30.55225 -->
<!--  1 	216.5		217.5			18.24349	 	25.65601	 	26.8512	 	28.95862	 	30.6114 -->
<!--  1 	217.5		218.5			18.28646	 	25.71481	 	26.91109	 	29.01921	 	30.67137 -->
<!--  1 	218.5		219.5			18.32904	 	25.77355	 	26.97108	 	29.0803	 	30.73222 -->
<!--  1 	219.5		220.5			18.37122	 	25.83225	 	27.03121	 	29.14191	 	30.794 -->
<!--  1 	220.5		221.5			18.41299	 	25.89093	 	27.09149	 	29.20409	 	30.85677 -->
<!--  1 	221.5		222.5			18.45432	 	25.94958	 	27.15194	 	29.26687	 	30.92058 -->
<!--  1 	222.5		223.5			18.4952	 	26.00823	 	27.21259	 	29.3303	 	30.9855 -->
<!--  1 	223.5		224.5			18.53562	 	26.06687	 	27.27344	 	29.39442	 	31.05158 -->
<!--  1 	224.5		225.5			18.57556	 	26.12553	 	27.33452	 	29.45926	 	31.11888 -->
<!--  1 	225.5		226.5			18.615	 	26.18422	 	27.39585	 	29.52487	 	31.18746 -->
<!--  1 	226.5		227.5			18.65393	 	26.24294	 	27.45746	 	29.5913	 	31.25739 -->
<!--  1 	227.5		228.5			18.69233	 	26.30171	 	27.51936	 	29.65857	 	31.32872 -->
<!--  1 	228.5		229.5			18.73019	 	26.36054	 	27.58159	 	29.72674	 	31.40152 -->
<!--  1 	229.5		230.5			18.76748	 	26.41945	 	27.64415	 	29.79585	 	31.47585 -->
<!--  1 	230.5		231.5			18.8042	 	26.47844	 	27.70707	 	29.86595	 	31.55178 -->
<!--  1 	231.5		232.5			18.84031	 	26.53753	 	27.77039	 	29.93707	 	31.62937 -->
<!--  1 	232.5		233.5			18.87581	 	26.59675	 	27.83411	 	30.00927	 	31.70868 -->
<!--  1 	233.5		234.5			18.91068	 	26.65609	 	27.89828	 	30.08258	 	31.78979 -->
<!--  1 	234.5		235.5			18.94489	 	26.71558	 	27.9629	 	30.15706	 	31.87275 -->
<!--  1 	235.5		236.5			18.97844	 	26.77522	 	28.02801	 	30.23276	 	31.95764 -->
<!--  1 	236.5		237.5			19.01129	 	26.83505	 	28.09363	 	30.30971	 	32.04453 -->
<!--  1 	237.5		238.5			19.04343	 	26.89507	 	28.15978	 	30.38797	 	32.13348 -->
<!--  1 	238.5		239.5			19.07484	 	26.9553	 	28.2265	 	30.46758	 	32.22457 -->
<!--  1 	239.5		240			19.10551	 	27.01575	 	28.29381	 	30.54859	 	32.31787 -->
<!--  1 	240		240.5			19.12055	 	27.04607	 	28.3277	 	30.58964	 	32.36537 -->
<!--  2 	24		24.5			14.39787	 	18.01821	 	18.44139	 	19.10624	 	19.56411 -->
<!--  2 	24.5		25.5			14.38019	 	17.97371	 	18.39526	 	19.05824	 	19.51534 -->
<!--  2 	25.5		26.5			14.34527	 	17.88749	 	18.30611	 	18.96595	 	19.42198 -->
<!--  2 	26.5		27.5			14.31097	 	17.80489	 	18.22103	 	18.87853	 	19.3341 -->
<!--  2 	27.5		28.5			14.27728	 	17.72586	 	18.13997	 	18.79591	 	19.25163 -->
<!--  2 	28.5		29.5			14.2442	 	17.65035	 	18.06285	 	18.718	 	19.17448 -->
<!--  2 	29.5		30.5			14.21175	 	17.5783	 	17.98962	 	18.64472	 	19.10255 -->
<!--  2 	30.5		31.5			14.17992	 	17.50965	 	17.92019	 	18.57599	 	19.03578 -->
<!--  2 	31.5		32.5			14.14871	 	17.44435	 	17.85452	 	18.51173	 	18.97407 -->
<!--  2 	32.5		33.5			14.11813	 	17.38235	 	17.79253	 	18.45187	 	18.91733 -->
<!--  2 	33.5		34.5			14.08818	 	17.3236	 	17.73416	 	18.39632	 	18.86548 -->
<!--  2 	34.5		35.5			14.05885	 	17.26804	 	17.67936	 	18.345	 	18.81843 -->
<!--  2 	35.5		36.5			14.03016	 	17.21564	 	17.62805	 	18.29784	 	18.77609 -->
<!--  2 	36.5		37.5			14.00209	 	17.16634	 	17.58019	 	18.25475	 	18.73838 -->
<!--  2 	37.5		38.5			13.97466	 	17.12009	 	17.53571	 	18.21567	 	18.7052 -->
<!--  2 	38.5		39.5			13.94786	 	17.07685	 	17.49455	 	18.18051	 	18.67647 -->
<!--  2 	39.5		40.5			13.92169	 	17.03658	 	17.45667	 	18.14919	 	18.6521 -->
<!--  2 	40.5		41.5			13.89615	 	16.99923	 	17.422	 	18.12165	 	18.632 -->
<!--  2 	41.5		42.5			13.87124	 	16.96476	 	17.39049	 	18.09781	 	18.61608 -->
<!--  2 	42.5		43.5			13.84697	 	16.93312	 	17.3621	 	18.07759	 	18.60425 -->
<!--  2 	43.5		44.5			13.82333	 	16.90428	 	17.33676	 	18.06093	 	18.59643 -->
<!--  2 	44.5		45.5			13.80033	 	16.8782	 	17.31442	 	18.04775	 	18.59253 -->
<!--  2 	45.5		46.5			13.77796	 	16.85483	 	17.29505	 	18.03799	 	18.59246 -->
<!--  2 	46.5		47.5			13.75624	 	16.83413	 	17.27858	 	18.03158	 	18.59614 -->
<!--  2 	47.5		48.5			13.73516	 	16.81606	 	17.26497	 	18.02844	 	18.60348 -->
<!--  2 	48.5		49.5			13.71472	 	16.80058	 	17.25417	 	18.02851	 	18.61441 -->
<!--  2 	49.5		50.5			13.69493	 	16.78765	 	17.24613	 	18.03174	 	18.62883 -->
<!--  2 	50.5		51.5			13.67579	 	16.77723	 	17.24081	 	18.03805	 	18.64667 -->
<!--  2 	51.5		52.5			13.65731	 	16.76927	 	17.23815	 	18.04738	 	18.66785 -->
<!--  2 	52.5		53.5			13.63948	 	16.76375	 	17.23811	 	18.05967	 	18.69229 -->
<!--  2 	53.5		54.5			13.62231	 	16.7606	 	17.24065	 	18.07486	 	18.71992 -->
<!--  2 	54.5		55.5			13.6058	 	16.75981	 	17.24571	 	18.09289	 	18.75065 -->
<!--  2 	55.5		56.5			13.58997	 	16.76132	 	17.25326	 	18.1137	 	18.78441 -->
<!--  2 	56.5		57.5			13.5748	 	16.76509	 	17.26324	 	18.13722	 	18.82113 -->
<!--  2 	57.5		58.5			13.56031	 	16.77108	 	17.2756	 	18.16341	 	18.86074 -->
<!--  2 	58.5		59.5			13.54649	 	16.77925	 	17.29031	 	18.19221	 	18.90315 -->
<!--  2 	59.5		60.5			13.53336	 	16.78956	 	17.30732	 	18.22355	 	18.94832 -->
<!--  2 	60.5		61.5			13.52091	 	16.80197	 	17.32657	 	18.25738	 	18.99615 -->
<!--  2 	61.5		62.5			13.50915	 	16.81644	 	17.34803	 	18.29365	 	19.04659 -->
<!--  2 	62.5		63.5			13.49808	 	16.83292	 	17.37165	 	18.3323	 	19.09957 -->
<!--  2 	63.5		64.5			13.4877	 	16.85138	 	17.39739	 	18.37327	 	19.15501 -->
<!--  2 	64.5		65.5			13.47802	 	16.87177	 	17.42519	 	18.41651	 	19.21286 -->
<!--  2 	65.5		66.5			13.46903	 	16.89405	 	17.45502	 	18.46197	 	19.27305 -->
<!--  2 	66.5		67.5			13.46075	 	16.91819	 	17.48683	 	18.50959	 	19.33552 -->
<!--  2 	67.5		68.5			13.45317	 	16.94415	 	17.52057	 	18.55932	 	19.4002 -->
<!--  2 	68.5		69.5			13.4463	 	16.97187	 	17.5562	 	18.61111	 	19.46703 -->
<!--  2 	69.5		70.5			13.44013	 	17.00134	 	17.59369	 	18.6649	 	19.53594 -->
<!--  2 	70.5		71.5			13.43467	 	17.03249	 	17.63297	 	18.72064	 	19.60689 -->
<!--  2 	71.5		72.5			13.42991	 	17.06531	 	17.67402	 	18.77829	 	19.6798 -->
<!--  2 	72.5		73.5			13.42587	 	17.09974	 	17.71678	 	18.83778	 	19.75462 -->
<!--  2 	73.5		74.5			13.42254	 	17.13575	 	17.76122	 	18.89907	 	19.83129 -->
<!--  2 	74.5		75.5			13.41992	 	17.17331	 	17.8073	 	18.96211	 	19.90976 -->
<!--  2 	75.5		76.5			13.41801	 	17.21237	 	17.85496	 	19.02685	 	19.98995 -->
<!--  2 	76.5		77.5			13.41681	 	17.2529	 	17.90417	 	19.09324	 	20.07183 -->
<!--  2 	77.5		78.5			13.41632	 	17.29485	 	17.95489	 	19.16123	 	20.15533 -->
<!--  2 	78.5		79.5			13.41654	 	17.3382	 	18.00708	 	19.23077	 	20.2404 -->
<!--  2 	79.5		80.5			13.41748	 	17.38291	 	18.06069	 	19.30182	 	20.32698 -->
<!--  2 	80.5		81.5			13.41912	 	17.42894	 	18.11569	 	19.37432	 	20.41502 -->
<!--  2 	81.5		82.5			13.42147	 	17.47626	 	18.17203	 	19.44822	 	20.50447 -->
<!--  2 	82.5		83.5			13.42453	 	17.52482	 	18.22968	 	19.52349	 	20.59528 -->
<!--  2 	83.5		84.5			13.42829	 	17.5746	 	18.28859	 	19.60008	 	20.68739 -->
<!--  2 	84.5		85.5			13.43276	 	17.62557	 	18.34873	 	19.67794	 	20.78075 -->
<!--  2 	85.5		86.5			13.43793	 	17.67768	 	18.41007	 	19.75702	 	20.87531 -->
<!--  2 	86.5		87.5			13.4438	 	17.7309	 	18.47255	 	19.83728	 	20.97103 -->
<!--  2 	87.5		88.5			13.45037	 	17.7852	 	18.53615	 	19.91867	 	21.06786 -->
<!--  2 	88.5		89.5			13.45764	 	17.84055	 	18.60082	 	20.00116	 	21.16573 -->
<!--  2 	89.5		90.5			13.4656	 	17.89692	 	18.66653	 	20.08469	 	21.26462 -->
<!--  2 	90.5		91.5			13.47425	 	17.95426	 	18.73325	 	20.16923	 	21.36447 -->
<!--  2 	91.5		92.5			13.48359	 	18.01256	 	18.80093	 	20.25473	 	21.46524 -->
<!--  2 	92.5		93.5			13.49362	 	18.07177	 	18.86955	 	20.34116	 	21.56688 -->
<!--  2 	93.5		94.5			13.50432	 	18.13187	 	18.93906	 	20.42846	 	21.66935 -->
<!--  2 	94.5		95.5			13.51571	 	18.19283	 	19.00943	 	20.51661	 	21.77259 -->
<!--  2 	95.5		96.5			13.52777	 	18.2546	 	19.08063	 	20.60555	 	21.87658 -->
<!--  2 	96.5		97.5			13.5405	 	18.31718	 	19.15262	 	20.69525	 	21.98126 -->
<!--  2 	97.5		98.5			13.5539	 	18.38051	 	19.22537	 	20.78568	 	22.0866 -->
<!--  2 	98.5		99.5			13.56797	 	18.44458	 	19.29884	 	20.87678	 	22.19255 -->
<!--  2 	99.5		100.5			13.58269	 	18.50936	 	19.37301	 	20.96853	 	22.29907 -->
<!--  2 	100.5		101.5			13.59807	 	18.57481	 	19.44784	 	21.06089	 	22.40613 -->
<!--  2 	101.5		102.5			13.6141	 	18.64091	 	19.52329	 	21.15381	 	22.51367 -->
<!--  2 	102.5		103.5			13.63077	 	18.70762	 	19.59935	 	21.24727	 	22.62168 -->
<!--  2 	103.5		104.5			13.64809	 	18.77493	 	19.67596	 	21.34123	 	22.73009 -->
<!--  2 	104.5		105.5			13.66605	 	18.8428	 	19.75312	 	21.43565	 	22.83889 -->
<!--  2 	105.5		106.5			13.68463	 	18.91121	 	19.83077	 	21.53049	 	22.94803 -->
<!--  2 	106.5		107.5			13.70384	 	18.98012	 	19.9089	 	21.62573	 	23.05747 -->
<!--  2 	107.5		108.5			13.72368	 	19.04952	 	19.98748	 	21.72133	 	23.16719 -->
<!--  2 	108.5		109.5			13.74413	 	19.11937	 	20.06647	 	21.81725	 	23.27714 -->
<!--  2 	109.5		110.5			13.76519	 	19.18965	 	20.14584	 	21.91347	 	23.3873 -->
<!--  2 	110.5		111.5			13.78685	 	19.26034	 	20.22558	 	22.00996	 	23.49762 -->
<!--  2 	111.5		112.5			13.80911	 	19.3314	 	20.30564	 	22.10667	 	23.60808 -->
<!--  2 	112.5		113.5			13.83197	 	19.40282	 	20.38601	 	22.20358	 	23.71865 -->
<!--  2 	113.5		114.5			13.85541	 	19.47457	 	20.46665	 	22.30066	 	23.82929 -->
<!--  2 	114.5		115.5			13.87943	 	19.54662	 	20.54754	 	22.39789	 	23.93997 -->
<!--  2 	115.5		116.5			13.90402	 	19.61895	 	20.62866	 	22.49522	 	24.05066 -->
<!--  2 	116.5		117.5			13.92918	 	19.69154	 	20.70997	 	22.59264	 	24.16134 -->
<!--  2 	117.5		118.5			13.9549	 	19.76436	 	20.79145	 	22.69011	 	24.27198 -->
<!--  2 	118.5		119.5			13.98118	 	19.83739	 	20.87308	 	22.78761	 	24.38254 -->
<!--  2 	119.5		120.5			14.008	 	19.91061	 	20.95484	 	22.88511	 	24.49299 -->
<!--  2 	120.5		121.5			14.03535	 	19.984	 	21.03669	 	22.98258	 	24.60333 -->
<!--  2 	121.5		122.5			14.06324	 	20.05753	 	21.11861	 	23.08	 	24.71351 -->
<!--  2 	122.5		123.5			14.09166	 	20.13118	 	21.20059	 	23.17734	 	24.82351 -->
<!--  2 	123.5		124.5			14.12059	 	20.20493	 	21.28259	 	23.27458	 	24.93331 -->
<!--  2 	124.5		125.5			14.15003	 	20.27876	 	21.3646	 	23.3717	 	25.04288 -->
<!--  2 	125.5		126.5			14.17997	 	20.35264	 	21.44659	 	23.46867	 	25.15221 -->
<!--  2 	126.5		127.5			14.21041	 	20.42657	 	21.52854	 	23.56546	 	25.26126 -->
<!--  2 	127.5		128.5			14.24133	 	20.50052	 	21.61043	 	23.66206	 	25.37002 -->
<!--  2 	128.5		129.5			14.27272	 	20.57446	 	21.69224	 	23.75845	 	25.47846 -->
<!--  2 	129.5		130.5			14.30459	 	20.64838	 	21.77396	 	23.8546	 	25.58657 -->
<!--  2 	130.5		131.5			14.33691	 	20.72227	 	21.85555	 	23.95049	 	25.69432 -->
<!--  2 	131.5		132.5			14.36969	 	20.79609	 	21.937	 	24.0461	 	25.80169 -->
<!--  2 	132.5		133.5			14.4029	 	20.86984	 	22.01829	 	24.14141	 	25.90868 -->
<!--  2 	133.5		134.5			14.43656	 	20.94349	 	22.0994	 	24.23641	 	26.01525 -->
<!--  2 	134.5		135.5			14.47063	 	21.01703	 	22.18031	 	24.33108	 	26.12139 -->
<!--  2 	135.5		136.5			14.50512	 	21.09045	 	22.26101	 	24.42539	 	26.22709 -->
<!--  2 	136.5		137.5			14.54002	 	21.16371	 	22.34148	 	24.51933	 	26.33233 -->
<!--  2 	137.5		138.5			14.57531	 	21.23681	 	22.4217	 	24.61288	 	26.43709 -->
<!--  2 	138.5		139.5			14.61099	 	21.30974	 	22.50166	 	24.70603	 	26.54136 -->
<!--  2 	139.5		140.5			14.64705	 	21.38246	 	22.58133	 	24.79876	 	26.64513 -->
<!--  2 	140.5		141.5			14.68347	 	21.45498	 	22.66071	 	24.89106	 	26.74838 -->
<!--  2 	141.5		142.5			14.72025	 	21.52727	 	22.73977	 	24.98291	 	26.8511 -->
<!--  2 	142.5		143.5			14.75737	 	21.59931	 	22.8185	 	25.0743	 	26.95328 -->
<!--  2 	143.5		144.5			14.79484	 	21.67111	 	22.89689	 	25.16522	 	27.0549 -->
<!--  2 	144.5		145.5			14.83262	 	21.74263	 	22.97493	 	25.25564	 	27.15596 -->
<!--  2 	145.5		146.5			14.87073	 	21.81386	 	23.05259	 	25.34557	 	27.25645 -->
<!--  2 	146.5		147.5			14.90914	 	21.8848	 	23.12987	 	25.43498	 	27.35636 -->
<!--  2 	147.5		148.5			14.94784	 	21.95543	 	23.20675	 	25.52387	 	27.45567 -->
<!--  2 	148.5		149.5			14.98682	 	22.02573	 	23.28323	 	25.61223	 	27.55439 -->
<!--  2 	149.5		150.5			15.02607	 	22.0957	 	23.35928	 	25.70005	 	27.6525 -->
<!--  2 	150.5		151.5			15.06559	 	22.16532	 	23.43491	 	25.78731	 	27.75 -->
<!--  2 	151.5		152.5			15.10535	 	22.23458	 	23.51008	 	25.87401	 	27.84688 -->
<!--  2 	152.5		153.5			15.14535	 	22.30346	 	23.58481	 	25.96013	 	27.94314 -->
<!--  2 	153.5		154.5			15.18558	 	22.37196	 	23.65907	 	26.04568	 	28.03877 -->
<!--  2 	154.5		155.5			15.22602	 	22.44007	 	23.73285	 	26.13065	 	28.13377 -->
<!--  2 	155.5		156.5			15.26666	 	22.50777	 	23.80615	 	26.21502	 	28.22813 -->
<!--  2 	156.5		157.5			15.30749	 	22.57506	 	23.87895	 	26.2988	 	28.32185 -->
<!--  2 	157.5		158.5			15.34849	 	22.64192	 	23.95126	 	26.38197	 	28.41494 -->
<!--  2 	158.5		159.5			15.38966	 	22.70835	 	24.02305	 	26.46453	 	28.50739 -->
<!--  2 	159.5		160.5			15.43098	 	22.77434	 	24.09433	 	26.54648	 	28.59919 -->
<!--  2 	160.5		161.5			15.47244	 	22.83987	 	24.16508	 	26.62782	 	28.69036 -->
<!--  2 	161.5		162.5			15.51403	 	22.90494	 	24.23529	 	26.70853	 	28.78088 -->
<!--  2 	162.5		163.5			15.55572	 	22.96954	 	24.30497	 	26.78862	 	28.87077 -->
<!--  2 	163.5		164.5			15.59752	 	23.03366	 	24.37411	 	26.86808	 	28.96002 -->
<!--  2 	164.5		165.5			15.63941	 	23.09731	 	24.44269	 	26.94692	 	29.04864 -->
<!--  2 	165.5		166.5			15.68136	 	23.16045	 	24.51071	 	27.02513	 	29.13663 -->
<!--  2 	166.5		167.5			15.72338	 	23.22311	 	24.57818	 	27.1027	 	29.22399 -->
<!--  2 	167.5		168.5			15.76544	 	23.28525	 	24.64508	 	27.17965	 	29.31073 -->
<!--  2 	168.5		169.5			15.80753	 	23.34689	 	24.71141	 	27.25597	 	29.39686 -->
<!--  2 	169.5		170.5			15.84964	 	23.40801	 	24.77716	 	27.33167	 	29.48237 -->
<!--  2 	170.5		171.5			15.89175	 	23.46861	 	24.84234	 	27.40673	 	29.56729 -->
<!--  2 	171.5		172.5			15.93385	 	23.52868	 	24.90694	 	27.48118	 	29.6516 -->
<!--  2 	172.5		173.5			15.97592	 	23.58823	 	24.97096	 	27.555	 	29.73533 -->
<!--  2 	173.5		174.5			16.01795	 	23.64723	 	25.0344	 	27.6282	 	29.81848 -->
<!--  2 	174.5		175.5			16.05992	 	23.7057	 	25.09725	 	27.70079	 	29.90107 -->
<!--  2 	175.5		176.5			16.10183	 	23.76363	 	25.15951	 	27.77277	 	29.98309 -->
<!--  2 	176.5		177.5			16.14364	 	23.82101	 	25.22119	 	27.84414	 	30.06456 -->
<!--  2 	177.5		178.5			16.18536	 	23.87784	 	25.28228	 	27.91491	 	30.1455 -->
<!--  2 	178.5		179.5			16.22696	 	23.93412	 	25.34279	 	27.98509	 	30.22591 -->
<!--  2 	179.5		180.5			16.26842	 	23.98985	 	25.40271	 	28.05468	 	30.3058 -->
<!--  2 	180.5		181.5			16.30974	 	24.04503	 	25.46204	 	28.12369	 	30.3852 -->
<!--  2 	181.5		182.5			16.35089	 	24.09964	 	25.5208	 	28.19213	 	30.46411 -->
<!--  2 	182.5		183.5			16.39185	 	24.1537	 	25.57897	 	28.26	 	30.54255 -->
<!--  2 	183.5		184.5			16.43262	 	24.20721	 	25.63656	 	28.32732	 	30.62053 -->
<!--  2 	184.5		185.5			16.47318	 	24.26015	 	25.69357	 	28.39408	 	30.69807 -->
<!--  2 	185.5		186.5			16.51351	 	24.31254	 	25.75002	 	28.46031	 	30.77519 -->
<!--  2 	186.5		187.5			16.55358	 	24.36437	 	25.80589	 	28.52602	 	30.8519 -->
<!--  2 	187.5		188.5			16.5934	 	24.41564	 	25.8612	 	28.5912	 	30.92822 -->
<!--  2 	188.5		189.5			16.63293	 	24.46636	 	25.91595	 	28.65588	 	31.00417 -->
<!--  2 	189.5		190.5			16.67216	 	24.51653	 	25.97014	 	28.72007	 	31.07976 -->
<!--  2 	190.5		191.5			16.71107	 	24.56614	 	26.02379	 	28.78378	 	31.15502 -->
<!--  2 	191.5		192.5			16.74965	 	24.61521	 	26.07689	 	28.84702	 	31.22997 -->
<!--  2 	192.5		193.5			16.78787	 	24.66372	 	26.12945	 	28.90981	 	31.30462 -->
<!--  2 	193.5		194.5			16.82573	 	24.7117	 	26.18148	 	28.97215	 	31.379 -->
<!--  2 	194.5		195.5			16.8632	 	24.75913	 	26.23299	 	29.03407	 	31.45314 -->
<!--  2 	195.5		196.5			16.90025	 	24.80603	 	26.28399	 	29.09558	 	31.52704 -->
<!--  2 	196.5		197.5			16.93689	 	24.8524	 	26.33448	 	29.1567	 	31.60075 -->
<!--  2 	197.5		198.5			16.97308	 	24.89824	 	26.38446	 	29.21743	 	31.67427 -->
<!--  2 	198.5		199.5			17.0088	 	24.94356	 	26.43396	 	29.27781	 	31.74764 -->
<!--  2 	199.5		200.5			17.04404	 	24.98836	 	26.48298	 	29.33784	 	31.82088 -->
<!--  2 	200.5		201.5			17.07879	 	25.03265	 	26.53153	 	29.39755	 	31.89401 -->
<!--  2 	201.5		202.5			17.11301	 	25.07643	 	26.57962	 	29.45695	 	31.96706 -->
<!--  2 	202.5		203.5			17.14669	 	25.11972	 	26.62726	 	29.51606	 	32.04007 -->
<!--  2 	203.5		204.5			17.17981	 	25.16251	 	26.67447	 	29.57491	 	32.11305 -->
<!--  2 	204.5		205.5			17.21234	 	25.20482	 	26.72125	 	29.6335	 	32.18603 -->
<!--  2 	205.5		206.5			17.24429	 	25.24665	 	26.76761	 	29.69187	 	32.25905 -->
<!--  2 	206.5		207.5			17.2756	 	25.28802	 	26.81358	 	29.75004	 	32.33212 -->
<!--  2 	207.5		208.5			17.30628	 	25.32892	 	26.85915	 	29.80802	 	32.40529 -->
<!--  2 	208.5		209.5			17.3363	 	25.36937	 	26.90436	 	29.86584	 	32.47859 -->
<!--  2 	209.5		210.5			17.36564	 	25.40938	 	26.9492	 	29.92352	 	32.55204 -->
<!--  2 	210.5		211.5			17.39427	 	25.44895	 	26.9937	 	29.98109	 	32.62567 -->
<!--  2 	211.5		212.5			17.42218	 	25.4881	 	27.03787	 	30.03857	 	32.69952 -->
<!--  2 	212.5		213.5			17.44935	 	25.52684	 	27.08173	 	30.09599	 	32.77362 -->
<!--  2 	213.5		214.5			17.47576	 	25.56517	 	27.12528	 	30.15337	 	32.84802 -->
<!--  2 	214.5		215.5			17.50137	 	25.60311	 	27.16856	 	30.21074	 	32.92272 -->
<!--  2 	215.5		216.5			17.52618	 	25.64067	 	27.21157	 	30.26812	 	32.99779 -->
<!--  2 	216.5		217.5			17.55015	 	25.67786	 	27.25433	 	30.32554	 	33.07324 -->
<!--  2 	217.5		218.5			17.57328	 	25.7147	 	27.29686	 	30.38304	 	33.14912 -->
<!--  2 	218.5		219.5			17.59553	 	25.75118	 	27.33918	 	30.44063	 	33.22546 -->
<!--  2 	219.5		220.5			17.61689	 	25.78733	 	27.3813	 	30.49835	 	33.30231 -->
<!--  2 	220.5		221.5			17.63733	 	25.82317	 	27.42325	 	30.55623	 	33.37969 -->
<!--  2 	221.5		222.5			17.65683	 	25.85869	 	27.46505	 	30.6143	 	33.45766 -->
<!--  2 	222.5		223.5			17.67537	 	25.89392	 	27.50671	 	30.6726	 	33.53624 -->
<!--  2 	223.5		224.5			17.69293	 	25.92887	 	27.54826	 	30.73114	 	33.61548 -->
<!--  2 	224.5		225.5			17.70948	 	25.96356	 	27.58971	 	30.78997	 	33.69542 -->
<!--  2 	225.5		226.5			17.725	 	25.99799	 	27.63109	 	30.84911	 	33.77609 -->
<!--  2 	226.5		227.5			17.73946	 	26.03219	 	27.67242	 	30.90861	 	33.85756 -->
<!--  2 	227.5		228.5			17.75286	 	26.06617	 	27.71372	 	30.96849	 	33.93984 -->
<!--  2 	228.5		229.5			17.76515	 	26.09993	 	27.75502	 	31.0288	 	34.023 -->
<!--  2 	229.5		230.5			17.77632	 	26.13351	 	27.79633	 	31.08956	 	34.10707 -->
<!--  2 	230.5		231.5			17.78635	 	26.16692	 	27.83769	 	31.15082	 	34.1921 -->
<!--  2 	231.5		232.5			17.79521	 	26.20016	 	27.8791	 	31.21261	 	34.27814 -->
<!--  2 	232.5		233.5			17.80288	 	26.23326	 	27.92061	 	31.27496	 	34.36522 -->
<!--  2 	233.5		234.5			17.80934	 	26.26624	 	27.96223	 	31.33793	 	34.45341 -->
<!--  2 	234.5		235.5			17.81456	 	26.29911	 	28.00399	 	31.40154	 	34.54273 -->
<!--  2 	235.5		236.5			17.81852	 	26.33189	 	28.04591	 	31.46583	 	34.63326 -->
<!--  2 	236.5		237.5			17.82119	 	26.36459	 	28.08801	 	31.53085	 	34.72503 -->
<!--  2 	237.5		238.5			17.82256	 	26.39723	 	28.13034	 	31.59664	 	34.8181 -->
<!--  2 	238.5		239.5			17.82259	 	26.42984	 	28.17291	 	31.66324	 	34.9125 -->
<!--  2 	239.5		240			17.82127	 	26.46243	 	28.21574	 	31.73069	 	35.00831 -->
<!--  2 	240		240.5			17.82009	 	26.47872	 	28.23727	 	31.76474	 	35.05675 -->
<!--  2 	240.5		241			17.81856	 	26.49502	 	28.25888	 	31.79903	 	35.10556 -->
<!--  " -->
<!-- ``` -->


<!-- ```{r} -->
<!-- BMI_cutoff <- read.delim(text = BMI_criterion, sep = "") -->
<!-- ``` -->

<!-- ```{r} -->
<!-- head(BMI_cutoff) -->
<!-- ``` -->
<!-- ```{r} -->
<!-- BMI_cutoff_matrix <- as.matrix(BMI_cutoff) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- head(BMI_cutoff_matrix ) -->
<!-- ``` -->

<!-- how to combine null with non-null data? -->
<!-- SAS and R use different strategy when no condition is met SAS return missing while R return NULL -->
<!-- which is not compatible with other data type -->
<!-- ```{r} -->
<!-- ## problematic logic -->
<!-- f <- function(sex, age_in_month, bmi){ -->
<!--   # for month > 240 if they don't have information about bmi and age_in_month then set as missing -->
<!--   if (is.na(bmi) || is.na(age_in_month)) { -->
<!--     return(NA) -->
<!--   } -->


<!--   # classify BMI category for age > 240 month -->
<!--   if (age_in_month >= 240) { -->
<!--     if (bmi > 0 && bmi < 18.5){ -->
<!--       return('a:underweight') -->
<!--     } -->
<!--     else if (bmi >= 18.5 && bmi < 25) { -->
<!--       return('b:normal') -->
<!--     } -->
<!--     else if (bmi >= 25 && bmi < 30) { -->
<!--       return('c:overweight') -->
<!--     } -->
<!--     else if (bmi >= 30){ -->
<!--       return('d:obese') -->
<!--     } -->
<!--   } -->
<!--   # the SAS code jump over 240 -->
<!--   else if (age_in_month < 240) { -->
<!--     if(is.na(sex)) { -->
<!--       return(NA) -->
<!--     } -->
<!--     # classify BMI category for age <= 240 month -->
<!--     for (group in 1:nrow(BMI_cutoff_matrix)) { -->
<!--       if (age_in_month >= BMI_cutoff_matrix[group, 'month_lower_limit'] && -->
<!--           age_in_month < BMI_cutoff_matrix[group, 'month_upper_limit']  && sex == BMI_cutoff_matrix[group, 'sex']) { -->
<!--             if (bmi > 0 && bmi <= BMI_cutoff_matrix[group, 'P5']) { -->
<!--               return('a:underweight') -->
<!--             } -->
<!--             else if (bmi > BMI_cutoff_matrix[group, 'P5'] && bmi <= BMI_cutoff_matrix[group, 'P85']) { -->
<!--               return('b:normal') -->
<!--             } -->
<!--             else if (bmi > BMI_cutoff_matrix[group, 'P85'] && bmi <= BMI_cutoff_matrix[group, 'P95']) { -->
<!--               return('c:overweight') -->
<!--             } -->
<!--             else if (bmi > BMI_cutoff_matrix[group, 'P95']) { -->
<!--               return('d:obese') -->
<!--             } -->
<!--       } -->
<!--     } -->
<!--   } -->

<!--   # if none of those condition is satisfied then return missing -->
<!--   return(NA) -->
<!-- } -->

<!-- ``` -->



<!-- ```{r} -->
<!-- rawdata2 <- rawdata2 %>% -->
<!--   rowwise() %>% -->
<!--   mutate(obese_overweight = f(sex, age_in_month, bmi)) -->
<!-- ``` -->



















simplified version for calculate BMI for adult
```{r}
rawdata2 <- rawdata2 %>% 
  mutate(obese_overweight = case_when(
    age >= 18 && bmi < 18.5 ~ 'a:underweight',
    age >= 18 && bmi >= 18.5 && bmi < 25 ~ 'b:normal',
    age >= 18 && bmi >= 25 && bmi < 39 ~ 'c:overweight',
    age >= 18 && bmi >= 30 ~ 'd:obese'
  ))

freq(rawdata2$obese_overweight)
```









```{r}
# if obese_overweight="d:obese" then do; obesity="Yes";obesity01=1; end;
# else if obese_overweight~="" then do; obesity="No";obesity01=0; end;
# 
# if obese_overweight in ("d:obese","c:overweight") then do; overweight="Yes";overweight01=1; end;
# else if obese_overweight~="" then do; overweight="No";overweight01=0; end;
#  
# run;


# create binary outcome variable of overweight
rawdata2 <- rawdata2 %>% 
  rowwise() %>% 
  mutate(obesity01 = ifelse(is.na(obese_overweight), NA, 
                           ifelse(obese_overweight == 'd:obese', 1, 0)),
         overweight = ifelse(is.na(obese_overweight), NA,
                               ifelse(obese_overweight %in% c("d:obese","c:overweight"), 'Yes', 'No')),
         overweight01 = ifelse(is.na(obese_overweight), NA,
                               ifelse(obese_overweight %in% c("d:obese","c:overweight"), 1, 0))
        )
```




```{r}
head(rawdata2)
```

```{r}
freq(rawdata2$obese_overweight)
```


```{r}
save(rawdata2, file = 'data/rawdata2.RData')
```




<!-- Comparison with the SAS result -->
<!-- ```{r} -->
<!-- # library(haven) -->
<!-- # dat <- read_sas('data/dat.sas7bdat') -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # dat <- dat %>%  -->
<!-- #   rowwise() %>%  -->
<!-- #   mutate(obese_overweight = ifelse(obese_overweight == '', NA, obese_overweight)) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # dat[9451, c('BMI', 'sex', 'age_in_month', 'obese_overweight', 'P5', 'P85', 'P90', 'P95', 'P97')] -->
<!-- ``` -->
<!-- ```{r} -->
<!-- # rawdata2[9451, c('bmi', 'sex', 'age_in_month')] -->
<!-- ``` -->


<!-- SAS -->
<!-- Missing values are treated as numbers equal to minus infinity.  So, any number is greater than a missing value. -->
<!-- ```{r} -->
<!-- # which((is.na(dat$obese_overweight) == is.na(rawdata2$obese_overweight)) == FALSE) -->
<!-- ``` -->
