==> R/step0 prepare data/step3 calculate the rests.Rmd <==
---
title: "step3 calculate the rests"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(readxl)
library(summarytools)
library(crosstable)
```


```{r}
load('data/rawdata2.Rdata')
```

```{r}
proc_freq <- function(data, vars) {
  for (var in vars) {
    print(var)
    print(freq(x = data[[var]]))
  }
}
```


<!-- ```{r} -->
<!-- proc_freq(rawdata2, c('chronic_kidney_disease', 'chronic_obstructive_pulmonary', 'down_syndrome', 'solid_organ_transplant',                              'obesity01', 'pregnancy', 'heart_failure', 'coronary_artery_disease', 'cardiomyopathies', 'sickle_cell_disease', 'smoke_yn', 'type_2_diabetes_mellitus')) -->
<!-- ``` -->


# directly use individual comorbidities but n_comorbidity is still wanted
# rawdata3
# check: charlson comorbidity index
## not bug (?) but logic weird
```{r}
# data rawdata3; set rawdata2;
# n_comorbidity=sum(Chronic_Kidney_Disease,
# Chronic_Obstructive_Pulmonary,Down_Syndrome,     # bug? sum will ignore missing value
# Solid_Organ_Transplant,
# obesity01,pregnancy,
# Heart_Failure,Coronary_Artery_Disease, 
# Cardiomyopathies,
# Sickle_Cell_Disease,smoke_yn,Type_2_Diabetes_Mellitus);
# 
# N_Posi_Comorb=sum(Asthma,Cerebrovascular,Cystic_Fibrosis,
# Hypertension,Other_Immuno,Liver_Disease,Dementia,
# overweight01,Pulmonary_Fibrosis,Thalassemia,Type_1_Diabetes_Mellitus);

rawdata3 <- rawdata2 %>%
  rowwise() %>%
  mutate(n_comorbidity = sum(any_malignancy, hiv, renal, hemiplegia, diabetes_with_cc,
                           diabetes_without_cc, severe_liver, mild_liver, 
                           peptic_ulcer, rheumatic, chronic_pulmonary, cerebrovascular,
                           dementia, peripheral_vascular, heart_failure,
                           myocardial_infarction, obesity, cardiovascular,
                           smoking_infarction,
                           wheelchair,
                           dysphagia, na.rm = TRUE),
         charlson_index = (cardiovascular + myocardial_infarction + heart_failure + peripheral_vascular + cerebrovascular +
           dementia + chronic_pulmonary + rheumatic + peptic_ulcer + mild_liver + diabetes_without_cc) +
           (hemiplegia + renal + diabetes_with_cc) + 3 * (severe_liver) + 6 * (hiv),
         liver_disease = ifelse(severe_liver == 1 || mild_liver == 1, 1, 0),
         diabetes = ifelse(diabetes_with_cc == 1 || diabetes_without_cc == 1, 1, 0))

#         n_posi_comorb=sum(asthma,cerebrovascular,cystic_fibrosis,
# hypertension,other_immuno,liver_disease,dementia,
# overweight01,pulmonary_fibrosis,thalassemia,type_1_diabetes_mellitus, na.rm = TRUE)
```





```{r}
# if n_comorbidity>=4 then n_comorbidity_gp=">=4";
# else if n_comorbidity=3 then n_comorbidity_gp="3";
# else if n_comorbidity=2 then n_comorbidity_gp="2";
# else if n_comorbidity=1 then n_comorbidity_gp="1";
# else if n_comorbidity~=. then n_comorbidity_gp="0";
#
# if N_Posi_Comorb>=4 then N_Posi_Comorb_gp=">=4";
# else if N_Posi_Comorb=3 then N_Posi_Comorb_gp="3";
# else if N_Posi_Comorb=2 then N_Posi_Comorb_gp="2";
# else if N_Posi_Comorb=1 then N_Posi_Comorb_gp="1"; # bug 0 is not considered
# else if N_Posi_Comorb~=. then N_Posi_Comorb_gp="0";


# create categorized n_comorbidity variable and n_posi_comorb
rawdata3 <- rawdata3 %>%
  # rowwise() %>%
  mutate(n_comorbidity_gp = ifelse(n_comorbidity >= 4, '>=4',
                                        ifelse(n_comorbidity == 3, '3',
                                               ifelse(n_comorbidity == 2, '2',
                                                      ifelse(n_comorbidity == 1, '1', '0')
                                                      )
                                               )
                                        ),
                                 
         charlson_index_gp = ifelse(charlson_index >= 4, '>=4',
                                        ifelse(charlson_index == 3, '3',
                                               ifelse(charlson_index == 2, '2',
                                                      ifelse(charlson_index == 1, '1', '0')
                                                      )
                                               )
                                        ),
         )
```



## new method to calculate treatment variables
# definition of systemic treatment ?
#bug (have been fixed)
```{r}
# if MA_Cytotoxic_4w=1 or PS_Cytotoxic_4w=1 or PR_Cytotoxic_4w=1 then cytotoxic_4wk="Y";
# else cytotoxic_4wk="N";
#
# if MA_Non_Cytotoxic_4w=1 or PS_Non_Cytotoxic_4w=1 or PR_Non_Cytotoxic_4w=1 then Nocytotoxic_4wk="Y";
# else Nocytotoxic_4wk="N";
#


# if cytotoxic_4wk="Y" then cytotoxic4wk="a. cytotoxic  ";                         # extra whitespace


# else if Nocytotoxic_4wk="Y" then cytotoxic4wk="b. no cytotoxic";
# else cytotoxic4wk="c. no chemo drug";
#
# *combine cytotoxic drug;
# if cytotoxic4wk in ("a. cytotoxic","b. no cytotoxi") then chemo_4w=1;            # variable value not match
# else if cytotoxic4wk="c. no chemo dr" then chemo_4w=0;                           # variablke value not match (c. no chemo dr)
#
# run;


# rawdata3 <- rawdata3 %>%
#   rowwise() %>%
#   # if NA is encountered in condition statement, NA will be return directly
#   mutate(cytotoxic_4wk = ifelse(ma_cytotoxic_4w == 1 || ps_cytotoxic_4w == 1 || pr_cytotoxic_4w == 1, 'Y', 'N'),
#          nocytotoxic_4wk = ifelse(ma_non_cytotoxic_4w == 1 || ps_non_cytotoxic_4w == 1 || pr_non_cytotoxic_4w == 1, 'Y', 'N'),
#          # cytotoxic4wk = ifelse(cytotoxic_4wk == 'Y', 'a. cytotoxic',
#          #                       ifelse(nocytotoxic_4wk == 'Y', 'b. no cytotoxic', 'c. no chemo drug')),
#          # combine cytotoxic drug
#          chemo_4w = ifelse(cytotoxic_4wk == 'Y' || nocytotoxic_4wk == 'Y', 1,
#                            ifelse(cytotoxic_4wk == 'N' && cytotoxic_4wk == 'N', 0, NA))
# 
#          )


rawdata3 <- rawdata3 %>%
  rowwise() %>%
  # if NA is encountered in condition statement, NA will be return directly
  mutate(chemotherapy_1_month = ifelse(chemotherapy_diag_1_month == 1 || chemotherapy_proc_1_month == 1, 1, 0),
         chemotherapy_3_month = ifelse(chemotherapy_diag_3_month == 1 || chemotherapy_proc_3_month == 1, 1, 0),
         chemotherapy_6_month = ifelse(chemotherapy_diag_6_month == 1 || chemotherapy_proc_6_month == 1, 1, 0),
         
         immunotherapy_1_month = immunotherapy_diag_1_month,
         immunotherapy_3_month = immunotherapy_diag_3_month,
         immunotherapy_6_month = immunotherapy_diag_6_month,
         
         systemic_1_month = ifelse(chemotherapy_1_month == 1 || immunotherapy_1_month == 1, 1, 0),
         systemic_3_month = ifelse(chemotherapy_3_month == 1 || immunotherapy_3_month == 1, 1, 0),
         systemic_6_month = ifelse(chemotherapy_6_month == 1 || immunotherapy_6_month == 1, 1, 0),
         
         radiation_1_month = ifelse(radiation_diag_1_month == 1 || radiation_proc_1_month == 1, 1, 0),
         radiation_3_month = ifelse(radiation_diag_3_month == 1 || radiation_proc_3_month == 1, 1, 0),
         radiation_6_month = ifelse(radiation_diag_6_month == 1 || radiation_proc_6_month == 1, 1, 0),
         
         treatment_remdesivir = ifelse(treatment_adm_remdesivir == 1 || treatment_ps_lopinavir_or_ritonavir == 1, 1, 0),
         treatment_lopinavir_or_ritonavir = ifelse(treatment_ps_lopinavir_or_ritonavir == 1 || treatment_adm_lopinavir_or_ritonavir == 1, 1, 0),
         treatment_bamlanivimab = ifelse(treatment_ps_bamlanivimab == 1 || treatment_adm_bamlanivimab == 1 || treatment_bamlanivimab_proc, 1, 0),
         
         treatment_casirivimab = ifelse(treatment_adm_casirivimab == 1 || treatment_ps_casirivimab == 1 || treatment_casirivimab_proc == 1, 1, 0),
         treatment_tocilizumab = ifelse(treatment_adm_tocilizumab == 1 || treatment_ps_tocilizumab == 1 || treatment_tocilizumab_proc, 1, 0),
         treatment_hydroxychloroquine = ifelse(treatment_adm_hydroxychloroquine == 1 || treatment_ps_hydroxychloroquine == 1, 1, 0))

```



# rawdata4
```{r}
# data rawdata4;
# set rawdata3; 
# cancer_year=year(cancer_date);
# 
# age_at_cancer=(cancer_year-BIRTH_YR);
# 
# duration=(COVID_diag_date-cancer_date)/365.25;
# if duration>5 then survive_5yr_more=1;
# else if duration<=5 and duration>=0 then survive_5yr_more=0;
# if cancer=0 then survive_5yr_more=.;
# 
# *IF the cancer occur before covid, then need to be labeled and deleted later;
# if duration<0 and duration~=. then cancer_after_COVID=1;
# 
# if age_at_cancer>=0 and age_at_cancer<20 then age_cancer_gp="0-20  ";
# else if age_at_cancer>=20 and age_at_cancer<40 then age_cancer_gp="20-40";
# else if age_at_cancer>=40 and age_at_cancer<60 then age_cancer_gp="40-60";
# else if age_at_cancer>=60 and age_at_cancer<80 then age_cancer_gp="60-80";
# else if age_at_cancer>=80 then age_cancer_gp=">=80";
# else age_cancer_gp="miss";


rawdata4 <- rawdata3 %>% 
  # rowwise() %>% 
  mutate(first_cancer_year = as.numeric(format(`first_cancer_date`, "%Y"))
         )


```

```{r}
rawdata4 <- rawdata4 %>% 
  rowwise() %>% 
  mutate(
         age_at_cancer = (first_cancer_year - birth_yr),
         duration = as.numeric(covid_pcr_date - first_cancer_date)/365.25,

         survive_5yr_more = ifelse(duration > 5, 1,
                                   ifelse(duration >=0 && duration <=5, 0, NA)),
         # override survive_5yr_more with special condition
         survive_5yr_more = ifelse(cancer == 0, NA, survive_5yr_more),

         # if the cancer occur before covid, then need to be labeled and deleted later;
         cancer_after_covid = ifelse(duration < 0, 1, NA),


         )

```


# unify the age categorization standard
```{r}
rawdata4 <- rawdata4 %>% 
  # rowwise() %>% 
  mutate(age_cancer_gp = ifelse(age_at_cancer >= 0 && age_at_cancer < 18, '0-18', 
                                ifelse(age_at_cancer >= 18 && age_at_cancer < 50, '18-50', 
                                       ifelse(age_at_cancer >= 50 && age_at_cancer < 65, '50-65',
                                              ifelse(age_at_cancer >= 65 && age_at_cancer < 75, '65-75',
                                                     ifelse(age_at_cancer >= 75, '>=75', NA))))))
```


```{r}
# proc freq data=rawdata4;
# tables age_cancer_gp Radiation_4w;
# where cancer=1;
# run;

vars <- c('age_cancer_gp','radiation_diag_6_month')
proc_freq(rawdata4[rawdata4$cancer == 1, ], vars)

```





# rawdata5
## survival data not accurate
##bug (not fixed) 
## a constant needed be changed when use new data (check whether the data is the version of the database)!
Anytime you use > you should check whether it is not missing in sas
```{r}
# data rawdata5;
# set rawdata4;
# time_death=DATE_OF_DEATH-COVID_diag_date;
# if time_death>=-15 and time_death<0 then time_death=0;*If it is within 15 days then set it as 0 (because we do not know death date but assuming mid of the month), immediately die after covid;
# if time_death~=. and time_death<-15 then time_death=.; *if the value is way below 15, treat it as missing;
# end_study=mdy("01","28","2021");
# format end_study mmddyy10.;
#  
# if death=1 then T=time_death;
# else if death=0 then T=(end_study-COVID_diag_date);
# 
# *if T<0 then T=.; *for those whose COVID_diag_date is over 
# *For those die after 30 days, death will be changed to "0" (means no death);
# T_greater_30=0;
# if death=1 and time_death>30 then do; T_greater_30=1; death=0;end;                # missing death will always < 30
# 
# if T~=. then T_30=min(T,30);
# 
# run;

rawdata5 <- rawdata4 %>% 
  # rowwise() %>% 
  mutate(time_death = as.numeric(date_of_death - `covid diag date`),
         time_death = ifelse(time_death >= -15 && time_death < 0, 0, 
                             ifelse(time_death < -15, NA, time_death)),
         end_study = as.Date('2021-01-31')                                       
         ) %>% 
  rowwise() %>% 
  mutate(t = ifelse(death == 1, time_death, 
                    ifelse(death == 0, as.numeric(end_study - `covid diag date`))),
         t = ifelse(t < 0, NA, t),
         # t_greater_30 = ifelse(death == 1 && time_death > 30, 1, 0),
         death = ifelse(death == 1 && time_death > 30, 0, death),
         t_30 = min(t, 30))
```



## create death month instead of days
```{r}
library(lubridate)
rawdata5 <- rawdata5 %>% 
  rowwise() %>%
  mutate(month_to_death = 12 * (year(date_of_death) - year(`covid diag date`)) + month(date_of_death) - month(`covid diag date`),
         # if death happen before the month of diag then treated as missing
      
         # by convention month_to_death_censor = 1 means it is observed
         month_to_death_censor = ifelse(is.na(month_to_death), 0, 1),
         month_to_death = ifelse(month_to_death_censor == 0, 12 * (year(as.Date('2021/01/08')) - year(`covid diag date`)) + month(as.Date('2021/01/08')) - month(`covid diag date`),
                                 ifelse(month_to_death < 0, NA, month_to_death)
         ),
         death_on_month_0 = ifelse((month_to_death_censor == 1) && (month_to_death == 0), 1, 0),
         death30 = ifelse((month_to_death_censor == 1) && (month_to_death <= 0), 1, 0),
         death90 = ifelse((month_to_death_censor == 1) && (month_to_death <= 3), 1, 0)
  )

```

```{r}
# proc print data=rawdata5 (obs=20);
# var death DATE_OF_DEATH COVID_diag_date  time_death T T_30;
#  run;

head(rawdata5[, c('death', 'date_of_death', 'covid diag date', 'time_death', 't', 't_30')], 20)
```


## no need to calculate cancer type (?)
```{r}
rawdata6 <- rawdata5
```

<!-- # rawdata6 -->
<!-- ##typo! (has been fixed) -->
<!-- ```{r} -->
<!-- # data rawdata6; -->
<!-- # set rawdata5; -->
<!-- # cancer_2digit=substr(Cancer_dx_code,1,3); -->
<!-- #  -->
<!-- # if cancer_2digit -->
<!-- # in ("C10", "C11", "C12", "C13", "C14", "C15", "C16", "C17", "C18", "C19", "C20", -->
<!-- #     "C21", "C22", "C23", "C24", "C25", "C26", "C30", "C31", "C32", "C33", "C34", -->
<!-- #     "C37", "C38", "C39", "C40", "C41", "C43", "C45", "C46", "C47", "C48", "C49", -->
<!-- #     "C50", "C51", "C52", "C53", "C54", "C55", "C56", "C57", "C58", "C60", "C61", -->
<!-- #     "C62", "C63", "C64", "C65", "C66", "C67", "C68", "C69", "C70", "C71", "C72", -->
<!-- #     "C73", "C74", "C75" ) then cancer_type="a. solid    ";                         # unnecessary trailing whitespace -->
<!-- # else if cancer_2digit in ("C81", "C82", "C83", "C84","C85", "C88", "C90", "C91", -->
<!-- #   "C92", "C93", "C94", "C95", "C96") then cancer_type="b. liquid"; -->
<!-- # else if cancer_2digit="C76" then cancer_type="c.Multitype"; -->
<!-- # if Region="Other/Unknown" then Region="z: Other/Unknown"; -->
<!-- # run; -->
<!-- #   -->

<!-- solid_cancer_codes <- c("C10", "C11", "C12", "C13", "C14", "C15", "C16", "C17", "C18", "C19", "C20", -->
<!--     "C21", "C22", "C23", "C24", "C25", "C26", "C30", "C31", "C32", "C33", "C34", -->
<!--     "C37", "C38", "C39", "C40", "C41", "C43", "C45", "C46", "C47", "C48", "C49", -->
<!--     "C50", "C51", "C52", "C53", "C54", "C55", "C56", "C57", "C58", "C60", "C61", -->
<!--     "C62", "C63", "C64", "C65", "C66", "C67", "C68", "C69", "C70", "C71", "C72", -->
<!--     "C73", "C74", "C75" ) -->

<!-- liquid_cancer_codes <- c("C81", "C82", "C83", "C84","C85", "C88", "C90", "C91", -->
<!--    "C92", "C93", "C94", "C95", "C96") -->

<!-- rawdata6 <- rawdata5 %>%  -->
<!--   rowwise() %>%  -->
<!--   mutate(cancer_2digit = substr(cancer_dx_code, 1, 3), -->
<!--          cancer_type = ifelse(cancer_2digit %in% solid_cancer_codes, 'a. solid',  -->
<!--                               ifelse(cancer_2digit %in% liquid_cancer_codes, 'b. liquid', -->
<!--                                      ifelse(cancer_2digit == 'C76', 'c.Multitype', NA))), -->
<!--          ) -->

<!-- ``` -->

```{r}
# proc freq data=rawdata6;
# tables Cancer_dx_code cancer_2digit  ;
# where cancer=1;
# run;

# proc_freq(rawdata6[rawdata6$cancer == 1, ], c('cancer_dx_code', 'cancer_2digit'))
```


# rawdata7
## bug!!!
```{r}
# data rawdata7;
# set rawdata6;
# if death=0 and inpatient=0 and ICU=0 and Ventilation=0 then severity_level=0;
# else if death=0 and inpatient=1 and ICU=0 and Ventilation=0 then severity_level=1;
# else if death=1 then severity_level=3;
# else if death~=. and inpatient~=. and ICU~=. and
# Ventilation~=. then severity_level=2;
# 
# if death=. then delete; *in outcomes, only death has missing;
# if cancer_after_COVID=1 then delete;   # cancer_after_COVID only have missing and 1 two values (which is a weird logic)
# run;

rawdata7 <- rawdata6 %>% 
  rowwise() %>% 
  mutate(severity_level = ifelse(death == 0 && inpatient == 0 && icu == 0 && ventilation == 0, 0,
                                 ifelse(death == 0 && inpatient == 1 && icu == 0 && ventilation == 0, 1,
                                        ifelse(death == 1, 3, 2)))
         )
  # filter(!is.na(death)) %>% 
  # filter(is.na(cancer_after_covid) || !(cancer_after_covid == 1)) # code logic weird
```


```{r}
# convert some variable to factor with level order
rawdata7 <- rawdata7 %>% 
  mutate(severity_level = factor(severity_level, levels=c('0', '1', '2', '3')),
         age_gp = factor(age_gp, levels=c('0-18', '18-50', '50-65', '65-75', '>= 75')),
         n_comorbidity_gp = factor(n_comorbidity_gp, levels = c('0', '1', '2', '3', '>=4')),
         charlson_index_gp = factor(charlson_index_gp, levels = c('0', '1', '2', '3', '>=4'))
         )
```




# data clean
```{r}
# filter covid before 2020/02/01
# filter death before COVID-19
rawdata8 <- rawdata7 %>% 
  filter((month_to_death >= 0) & (covid_pcr_date >= as.Date('2020/02/01')))
nrow(rawdata8)
```


# exlusion

## restrict data used (june to dec)
```{r}
# need use vectorized and operator (&)
rawdata8_inclusion <- rawdata8 %>% filter((covid_pcr_date >= as.Date('2020/06/01')) & (covid_pcr_date <= as.Date('2020/12/31')))
nrow(rawdata8_inclusion)
```

# exclusion 1
```{r}
rawdata8_exclusion1 <- rawdata8_inclusion %>%
  rowwise() %>% 
  filter(!is.na(age) && age >= 18 && !is.na(gender))
nrow(rawdata8_exclusion1)
```


# exclusion 2
at least require patients to enter the database for 365 days 
question: to make sure enough time to make cancer effect to take effect? or to make sure cancer info to be caught?
-We decided to limit patients who had any prior record within year to ensure look back time to capture prior cancer treatment and comorbidities. Instead of limiting patients, can you add cohort entering and exit dates without dropping patients, so I can run sensitivity analysis as well?
-Cancer treatment in different time windows: I added different cancer treatment windows so we can run sensitivity analysis as well (1 month, 3 month, 6 month)
```{r}
rawdata8_exclusion2 <- rawdata8_exclusion1 %>% 
  filter(difftime(covid_pcr_date, entry_date, units='days') >= 365) 
nrow(rawdata8_exclusion2)
```


```{r}
proc_freq(rawdata8_exclusion2, 'cancer')
proc_freq(rawdata8_exclusion2[rawdata8_exclusion2$cancer == 1, ], 'recent_cancer')
```



```{r}
final_dat <- rawdata8_exclusion2
save(final_dat, file = 'data/final_dat.RData')
```
